Glific (Glific v4.3.0)

Glific keeps the contexts that define your domain and business logic.

Contexts are also responsible for managing your data, regardless if it comes from the database, an external API or others.

For now we'll keep some commonly used functions here, until we need a new file

Functions

atomize_keys(map)

@spec atomize_keys(any()) :: any()

Convert map string keys to :atom keys

delete_multiple(map, list)

@spec delete_multiple(map(), list()) :: map()

Delete multiple items from the map

execute_eex(content)

@spec execute_eex(String.t()) :: String.t()

execute string in eex

go_back_time(go_back, time \\ DateTime.utc_now(), unit \\ :hour)

@spec go_back_time(integer(), DateTime.t(), atom()) :: DateTime.t()

Return a time object where you go back x units. We introduce the notion of hour and minute

ignore_error?(error)

@spec ignore_error?(String.t()) :: boolean()

A hack to suppress error messages when running lots of flows. These are expected and we want to improve signal <-> noise ratio

in_past_time(time, units \\ :hours, back \\ 24)

@spec in_past_time(DateTime.t(), atom(), integer()) :: boolean()

See if the current time is within the past time units

log_error(error, send_appsignal? \\ true)

@spec log_error(String.t(), boolean()) :: {:error, String.t()}

Log the error and also send it over to our friends at appsignal

make_set(str, seperators \\ [",", ";"])

@spec make_set(String.t(), list()) :: MapSet.t()

Given a string seperated by spaces, commas, or semi-colons, create a set of individual elements in the string

parse_maybe_integer!(value)

@spec parse_maybe_integer!(String.t() | integer()) :: integer()

parse and integer and die if parse fails

parse_maybe_integer(value)

@spec parse_maybe_integer(String.t() | integer()) :: {:ok, integer()} | :error

Wrapper to return :ok/:error when parsing strings to potential integers

safe_string_to_atom(value)

@spec safe_string_to_atom(String.t() | atom()) :: atom()

You shouldn’t really use String.to_atom/1 on user-supplied data. The BEAM has a limit on how many different atoms you can have and they’re not garbage collected! With data coming from outside the system, stick to strings or use String.to_existing_atom/1 instead! So this is a generic function which will convert the string to atom and throws an error in case of invalid key

signature(organization_id, body, timestamp)

@spec signature(non_neg_integer(), String.t(), non_neg_integer()) :: String.t()

Compute the signature at a specific time for the body

stacktrace()

@spec stacktrace() :: String.t()

easy way for glific developers to get a stacktrace when debugging

string_clean(str)

@spec string_clean(String.t() | nil) :: String.t() | nil

Lets get rid of all non valid characters. We are assuming any language and hence using unicode syntax and not restricting ourselves to alphanumeric

substitute_organization_id(params, value, key)

@spec substitute_organization_id(map(), any(), atom()) :: map()

Intermediary function to update the input params with organization id as operation is performed by glific_admin for other organizations

suspicious_code(code)

@spec suspicious_code(String.t()) :: boolean()

Really simple function to ensure folks do not add Repo and/or IO calls to an EEx snippet. This is an extremely short term fix to avoid shooting ourselves in the foot, but we should move to lua for flows scripting in the near future

Note that folks can potentially find other ways to access the same modules, so this by no means should be considered remotely secure

validate_shortcode(changeset)

@spec validate_shortcode(Ecto.Changeset.t()) ::
  Ecto.Changeset.t() | Ecto.Changeset.t()

Validates inputed shortcode, if shortcode is invalid it returns message that the shortcode is invalid along with the valid shortcode.