Glific.AccessControl (Glific v4.6.8)

The AccessControl context.

Functions

check_access(entity_list, entity_type)

@spec check_access(Ecto.Query.t(), atom()) :: Ecto.Query.t()

Common function to filtering entity objects based on user role, fun_with_flag flag and entity type

count_roles(args)

@spec count_roles(map()) :: integer()

Return the count of roles, using the same filter as list_roles

create_permission(attrs \\ %{})

@spec create_permission(map()) ::
  {:ok, Glific.AccessControl.Permission.t()} | {:error, Ecto.Changeset.t()}

Creates a permission.

Examples

iex> create_permission(%{field: value})
{:ok, %Permission{}}

iex> create_permission(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_role(attrs \\ %{})

@spec create_role(map()) ::
  {:ok, Glific.AccessControl.Role.t()} | {:error, Ecto.Changeset.t()}

Creates a role.

Examples

iex> create_role(%{field: value})
{:ok, %Role{}}

iex> create_role(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_permission(permission)

@spec delete_permission(Glific.AccessControl.Permission.t()) ::
  {:ok, Glific.AccessControl.Permission.t()} | {:error, Ecto.Changeset.t()}

Deletes a permission.

Examples

iex> delete_permission(permission)
{:ok, %Permission{}}

iex> delete_permission(permission)
{:error, %Ecto.Changeset{}}

delete_role(role)

@spec delete_role(Glific.AccessControl.Role.t()) ::
  {:ok, Glific.AccessControl.Role.t()} | {:error, Ecto.Changeset.t()}

Deletes a role.

Examples

iex> delete_role(role)
{:ok, %Role{}}

iex> delete_role(role)
{:error, %Ecto.Changeset{}}

do_check_access(entity_list, entity_type, user)

@spec do_check_access(Ecto.Query.t(), atom(), Glific.Users.User.t()) ::
  Ecto.Query.t() | {:error, String.t()}

Common function to filtering entity objects based on user role, fun_with_flag flag and entity type

get_permission!(id)

@spec get_permission!(integer()) :: Glific.AccessControl.Permission.t()

Gets a single permission.

Raises Ecto.NoResultsError if the Permission does not exist.

Examples

iex> get_permission!(123)
%Permission{}

iex> get_permission!(456)
** (Ecto.NoResultsError)

get_role!(id)

@spec get_role!(integer()) :: Glific.AccessControl.Role.t()

Gets a single role.

Raises Ecto.NoResultsError if the Role does not exist.

Examples

iex> get_role!(123)
%Role{}

iex> get_role!(456)
** (Ecto.NoResultsError)

list_permissions()

@spec list_permissions() :: [Glific.AccessControl.Permission.t()]

Returns the list of permissions.

Examples

iex> list_permissions()
[%Permission{}, ...]

list_roles(args)

@spec list_roles(map()) :: [Glific.AccessControl.Role.t()]

Returns the list of roles. Checks if roles and permission flag is enabled first. If roles and permission is enabled then list all roles If roles and permission is disabled then list only default roles

Examples

iex> list_roles()
[%Role{}, ...]

organization_roles(args)

@spec organization_roles(map()) :: [Glific.AccessControl.Role.t()]

Returns the labels of organization roles.

Examples

iex> organization_roles()
["Teacher", "Mentor"]

update_permission(permission, attrs)

@spec update_permission(Glific.AccessControl.Permission.t(), map()) ::
  {:ok, Glific.AccessControl.Permission.t()} | {:error, Ecto.Changeset.t()}

Updates a permission.

Examples

iex> update_permission(permission, %{field: new_value})
{:ok, %Permission{}}

iex> update_permission(permission, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_role(role, attrs)

@spec update_role(Glific.AccessControl.Role.t(), map()) ::
  {:ok, Glific.AccessControl.Role.t()} | {:error, Ecto.Changeset.t()}

Updates a role.

Examples

iex> update_role(role, %{field: new_value})
{:ok, %Role{}}

iex> update_role(role, %{field: bad_value})
{:error, %Ecto.Changeset{}}