Glific.Bigquery (Glific v1.3.3)

Glific Bigquery Dataset and table creation

Functions

alter_tables(conn, dataset_id, project_id)

Specs

alter_tables(Tesla.Client.t(), String.t(), String.t()) :: :ok

Alter bigquery table schema, if required this function should be called from iex

do_refresh_the_schema(organization_id, cred)

Specs

do_refresh_the_schema(non_neg_integer(), map()) ::
  {:error, Tesla.Env.t()} | {:ok, Tesla.Env.t()}

Refresh the biquery schema and update all the older versions.

format_data_for_bigquery(data, table)

Specs

format_data_for_bigquery(map(), String.t()) :: map()

Format Data for bigquery error

format_date(date, organization_id)

Specs

format_date(DateTime.t() | nil, non_neg_integer()) :: String.t()

Format dates for the bigquery.

format_json(definition)

Specs

format_json(map()) :: iodata()

Format all the json values

make_insert_query(data, table, organization_id, max_id)

Specs

make_insert_query(
  map() | list(),
  String.t(),
  non_neg_integer(),
  non_neg_integer()
) :: :ok

Insert rows in the biqquery

make_job_to_remove_duplicate(table, organization_id)

Specs

make_job_to_remove_duplicate(String.t(), non_neg_integer()) :: :ok

Merge delta and main tables.

sync_schema_with_bigquery(organization_id)

Specs

sync_schema_with_bigquery(non_neg_integer()) :: :ok

Creating a dataset with messages and contacts as tables