Glific.Clients (Glific v1.1.5)

Wrapper module that allows us to invoke organization specific callback functions to tweak the way the system handles things. This allows clients to override functionality in a similar manner to plugins wordpress.

At some point we will move this to a more extensible scheme, which is as yet undetermined

Functions

blocked?(phone, organization_id)

Specs

blocked?(String.t(), non_neg_integer()) :: boolean()

Programmatially block clients based on organization needs. Use case could be: Allow only numbers from India and US

broadcast(action, contact, staff_id)

Specs

broadcast(
  Glific.Flows.Action.t(),
  Glific.Contacts.Contact.t(),
  non_neg_integer()
) :: non_neg_integer()

Allow an organization to dynamically select which contact the broadcast message should go to. This gives NGOs more flexibility when building flows

gcs_params(media, bucket)

Specs

gcs_params(map(), String.t()) :: {String.t(), String.t()}

Overwrite the default GCS storage bucket