Glific.Clients.ArogyaWorld (Glific v4.3.0)

Custom code extenison for ArogyaWorld

Functions

add_data_from_csv(key, file_url, cleanup_func, org_id)

@spec add_data_from_csv(String.t(), String.t(), any(), non_neg_integer()) ::
  {:ok, any()} | {:error, Ecto.Changeset.t()}

add data that needs to be sent to the database

add_weekly_dynamic_data(key, file_url, org_id)

@spec add_weekly_dynamic_data(String.t(), String.t(), non_neg_integer()) ::
  {:ok, any()} | {:error, Ecto.Changeset.t()}

adds the weekly dynamic data loaded from the sheet based on current week

cleanup_static_data(acc, data)

@spec cleanup_static_data(map(), map()) :: map()

Clean static weekly data from the CSV file.

cleanup_week_data(acc, data)

@spec cleanup_week_data(map(), map()) :: map()

Clean week data from the CSV file.

get_messages_by_flow_label(label)

@spec get_messages_by_flow_label(String.t()) :: any()

Get the messages based on flow label

get_question_id(label)

@spec get_question_id(String.t()) :: any()

Return the question id based on the label

get_response(list, contact_id)

@spec get_response(list(), String.t()) :: String.t() | nil

Return the response of the question for a contact

get_response_score(response, q_id, org_id)

@spec get_response_score(String.t(), String.t(), non_neg_integer()) :: any()

Return the response score based on the body

get_responses_by_week_and_day(org_id, week, day)

@spec get_responses_by_week_and_day(non_neg_integer(), non_neg_integer(), String.t()) ::
  any()

Get response message based on day and week

hourly_tasks(org_id)

@spec hourly_tasks(non_neg_integer()) :: any()

Hourly task jobs for the ArogyaWorld

initial_load(org_id)

@spec initial_load(non_neg_integer()) :: any()

Run this function on the initial load

load_participant_file(org_id, week_number)

@spec load_participant_file(non_neg_integer(), non_neg_integer()) :: any()

load participant files from gcs

message_hsm_mapping(file_url, org_id)

@spec message_hsm_mapping(String.t(), non_neg_integer()) ::
  {:ok, any()} | {:error, Ecto.Changeset.t()}

message mapping to HSM UUID

question_hsm_mapping(file_url, org_id)

@spec question_hsm_mapping(String.t(), non_neg_integer()) ::
  {:ok, any()} | {:error, Ecto.Changeset.t()}

question mapping to HSM UUID

response_score_mapping(file_url, org_id)

@spec response_score_mapping(String.t(), non_neg_integer()) ::
  {:ok, any()} | {:error, Ecto.Changeset.t()}

response to score mapping

static_message_schedule_map(file_url, org_id)

@spec static_message_schedule_map(String.t(), non_neg_integer()) ::
  {:ok, any()} | {:error, Ecto.Changeset.t()}

creates the static data map that needs to be sent to users

template(template_uuid)

@spec template(integer()) :: binary()

get template form EEx without variables

template(template_uuid, variables)

@spec template(integer(), String.t()) :: binary()

get template form EEx based on variables

trigger_condition(trigger)

@spec trigger_condition(Glific.Triggers.Trigger.t()) :: boolean()

Conditionally execute the trigger based on: ID, Week, Day.

upload_participant_responses(org_id, week)

@spec upload_participant_responses(non_neg_integer(), non_neg_integer()) :: any()

Create a file in GCS bucket for candidate response

webhook(arg1, fields)

@spec webhook(String.t(), map()) :: map()

Webhook functions for the ArogyaWorld