Glific.Clients.Tap (Glific v4.6.8)

Tweak GCS Bucket name based on group that the contact is in (if any)

Functions

fix_contact_name()

@spec fix_contact_name() :: :ok

Fix the contact name issue

gcs_file_name(media)

@spec gcs_file_name(map()) :: String.t()

In the case of TAP we retrive the first group the contact is in and store and set the remote name to be a sub-directory under that group (if one exists)

maybe_add_profile_activity(activity_info, contact_id, org_id)

@spec maybe_add_profile_activity(map(), non_neg_integer(), non_neg_integer()) :: map()

Check if a contact has more profiles and add that to message.

template(shortcode, params_staring \\ "")

@spec template(String.t(), String.t()) :: binary()

get template form EEx without variables

update_all_profile_fields(org_id, contact_id, key, value)

@spec update_all_profile_fields(
  non_neg_integer(),
  non_neg_integer(),
  String.t(),
  String.t()
) :: :ok

Update the fields for all the profiles.

webhook(arg1, fields)

@spec webhook(String.t(), map()) :: map()

Create a webhook with different signatures, so we can easily implement additional functionality as needed