Glific v0.1.0 Glific.Communications

Glific interface for all provider communication

Functions

provider()

Specs

provider() :: atom()

Get the current provider based on the config

provider_worker()

Specs

provider_worker() :: atom()

Get the current provider worker based on the organization | Config | Defaultconfig

publish_data(data, topic)

Specs

publish_data(
  {:ok, Glific.Messages.Message.t() | Glific.Tags.MessageTag.t()},
  atom()
) :: Glific.Messages.Message.t() | Glific.Tags.MessageTag.t()

publish_data(Glific.Messages.Message.t() | Glific.Tags.MessageTag.t(), atom()) ::
  Glific.Messages.Message.t() | Glific.Tags.MessageTag.t()

Unified function to publish data on the graphql subscription endpoint. This is still looking for a place to actually reside. This is a good next stop for now

For now the data types are Message and MessageTag