Glific.Communications.Message (Glific v0.9.4)

The Message Communication Context, which encapsulates and manages tags and the related join tables.

Functions

handle_error_response(response, message)

Specs

handle_error_response(Tesla.Env.t(), Glific.Messages.Message.t()) ::
  {:error, String.t()}

Callback in case of any error while sending the message

handle_success_response(response, message)

Specs

handle_success_response(Tesla.Env.t(), Glific.Messages.Message.t()) ::
  {:ok, Glific.Messages.Message.t()}

Callback when message send succsully

receive_message(message_params, type \\ :text)

Specs

receive_message(map(), atom()) :: {:ok} | {:error, String.t()}

Callback when we receive a message from whats app

send_message(message, attrs \\ %{})

Specs

send_message(Glific.Messages.Message.t(), map()) ::
  {:ok, Glific.Messages.Message.t()} | {:error, String.t()}

Send message to receiver using define provider.

update_bsp_status(bsp_message_id, bsp_status, errors)

Specs

update_bsp_status(String.t(), atom(), map()) :: any()

Callback to update the provider status for a message