Glific v0.1.0 Glific.Conversations

The main Glific abstraction that exposes the data in a stuctured manner as a set of conversations. For now each contact is associated with one and only one conversation. We will keep the API simple for now, but it is likely to become more complex and will require a fair number of iterations to get right

Functions

conversation_by_id(args)

Specs

conversation_by_id(map()) :: Glific.Conversations.Conversation.t() | nil

Returns the filtered conversation by contact id

list_conversations(args)

Specs

list_conversations(map()) :: list()

Returns the last M conversations, each conversation not more than N messages