Glific v0.3.1 Glific.Dialogflow.Intents

The main intents module which stiches it all together

Functions

add_training_phrase(id, text, language \\ "en")

Specs

add_training_phrase(String.t(), String.t(), String.t()) :: tuple()

Add a training phrase to an intent.

create(body, language \\ "es")

Specs

create(map(), String.t()) :: tuple()

Create an intent

get(id, language \\ "en", view \\ "INTENT_VIEW_UNSPECIFIED")

Specs

get(String.t(), String.t(), String.t()) :: tuple()

Get an intent for a specific id

list(language \\ "en", view \\ "INTENT_VIEW_UNSPECIFIED", token \\ nil, acc \\ [])

Specs

list(String.t(), String.t(), String.t() | nil, list()) :: tuple()

List all the intents of the agent per pageToken.

list_by_page(language \\ "en", view \\ "INTENT_VIEW_UNSPECIFIED", page_size \\ 100, token \\ nil)

Specs

list_by_page(String.t(), String.t(), integer(), String.t() | nil) :: tuple()

List an agent's intents by pageToken and defining page size.

update(id, intent, language \\ "es")

Specs

update(String.t(), map(), String.t()) :: tuple()

Update an intent view full.