Glific.Enums (Glific v4.3.0)

The Enum provides a location for all enum related macros. All the constants that Ecto/Elixir used are exposed here as macros, so other files can invoke them as simple functions

Functions

api_status_const()
(macro)

Lets ensure that the constants evaluate to some expected values

Examples

iex> Glific.Enums.api_status_const() [:ok, :error]

iex> Glific.Enums.contact_status_const() [:blocked, :failed, :invalid, :processing, :valid]

iex> Glific.Enums.contact_provider_status_const() [:none, :session, :session_and_hsm, :hsm]

iex> Glific.Enums.flow_case_const() [:has_any_word]

iex> Glific.Enums.flow_router_const() [:switch]

iex> Glific.Enums.flow_action_type_const() [:enter_flow, :send_msg, :set_contact_language, :wait_for_response, :set_contact_field]

iex> Glific.Enums.flow_type_const() [:message]

iex> Glific.Enums.message_flow_const() [:inbound, :outbound]

iex> Glific.Enums.message_status_const() [:sent, :delivered, :enqueued, :error, :read, :received, :contact_opt_out]

iex> Glific.Enums.message_type_const() [:audio, :contact, :document, :hsm, :image, :location, :list, :quick_reply, :text, :video, :sticker]

iex> Glific.Enums.question_type_const() [:text, :numeric, :date]

iex> Glific.Enums.sort_order_const() [:asc, :desc]

We also test the ecto enums in this file, since they exist outside a module

iex> Glific.Enums.APIStatus.enum_map() Glific.Enums.api_status_const()

iex> Glific.Enums.ContactStatus.enum_map() Glific.Enums.contact_status_const()

iex> Glific.Enums.ContactProviderStatus.enum_map() Glific.Enums.contact_provider_status_const()

iex> Glific.Enums.FlowCase.enum_map() Glific.Enums.flow_case_const()

iex> Glific.Enums.FlowRouter.enum_map() Glific.Enums.flow_router_const()

iex> Glific.Enums.FlowActionType.enum_map() Glific.Enums.flow_action_type_const()

iex> Glific.Enums.FlowType.enum_map() Glific.Enums.flow_type_const()

iex> Glific.Enums.MessageFlow.enum_map() Glific.Enums.message_flow_const()

iex> Glific.Enums.MessageStatus.enum_map() Glific.Enums.message_status_const()

iex> Glific.Enums.MessageType.enum_map() Glific.Enums.message_type_const()

iex> Glific.Enums.QuestionType.enum_map() Glific.Enums.question_type_const()

iex> Glific.Enums.SortOrder.enum_map() Glific.Enums.sort_order_const()

iex> Glific.Enums.ContactFieldValueType.enum_map() Glific.Enums.contact_field_value_type_const()

iex> Glific.Enums.ContactFieldScope.enum_map() Glific.Enums.contact_field_scope_const()

iex> Glific.Enums.UserRoles.enum_map() Glific.Enums.user_roles_const()

iex> Glific.Enums.TemplateButtonType.enum_map() Glific.Enums.template_button_type_const()

iex> Glific.Enums.OrganizationStatus.enum_map() Glific.Enums.organization_status_const()

iex> Glific.Enums.InteractiveMessageType.enum_map() Glific.Enums.interactive_message_type_const()

iex> Glific.Enums.ImportContactsType.enum_map() Glific.Enums.import_contacts_type_const()

contact_field_scope_const()
(macro)

contact_field_value_type_const()
(macro)

contact_provider_status_const()
(macro)

contact_status_const()
(macro)

flow_action_type_const()
(macro)

flow_case_const()
(macro)

flow_router_const()
(macro)

flow_type_const()
(macro)

import_contacts_type_const()
(macro)

interactive_message_type_const()
(macro)

message_flow_const()
(macro)

message_status_const()
(macro)

message_type_const()
(macro)

organization_status_const()
(macro)

question_type_const()
(macro)

sort_order_const()
(macro)

template_button_type_const()
(macro)

user_roles_const()
(macro)