Glific.Extensions (Glific v0.9.4)

The Extensions Context, which encapsulates and manages dynamic loading and unloading of elixir code

Functions

change_extension(extension, attrs \\ %{})

Specs

change_extension(Glific.Extensions.Extension.t(), map()) :: Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking extension changes.

Examples

iex> change_extension(extension)
%Ecto.Changeset{data: %Extension{}}

count_extensions(args)

Specs

count_extensions(map()) :: integer()

Return the count of extensions, using the same filter as list_extensions

create_extension(attrs)

Specs

create_extension(map()) ::
  {:ok, Glific.Extensions.Extension.t()} | {:error, Ecto.Changeset.t()}

Creates a extension.

Examples

iex> create_extension(%{field: value})
{:ok, %Extension{}}

iex> create_extension(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_extension(extension)

Specs

delete_extension(Glific.Extensions.Extension.t()) ::
  {:ok, Glific.Extensions.Extension.t()} | {:error, Ecto.Changeset.t()}

Deletes a extension.

Examples

iex> delete_extension(extension)
{:ok, %Extension{}}

iex> delete_extension(extension)
{:error, %Ecto.Changeset{}}

execute(name, body)

Specs

execute(String.t() | Glific.Extensions.Extension.t(), map()) :: map()

Given an extension name. It first requires the file and if that succeeds, it then calls the function in the specified module and returns the json result

get_extension!(id)

Specs

get_extension!(integer()) :: Glific.Extensions.Extension.t()

Gets a single extension.

Raises Ecto.NoResultsError if the Extension does not exist.

Examples

iex> get_extension!(123)
%Extension{}

iex> get_extension!(456)
** (Ecto.NoResultsError)

list_extensions(args)

Specs

list_extensions(map()) :: [Glific.Extensions.Extension.t()]

Returns the list of extensions.

Examples

iex> list_extensions()
[%Extension{}, ...]

update_extension(extension, attrs)

Specs

update_extension(Glific.Extensions.Extension.t(), map()) ::
  {:ok, Glific.Extensions.Extension.t()} | {:error, Ecto.Changeset.t()}

Updates a extension.

Examples

iex> update_extension(extension, %{field: new_value})
{:ok, %Extension{}}

iex> update_extension(extension, %{field: bad_value})
{:error, %Ecto.Changeset{}}