Glific.Flows (Glific v4.3.0)

The Flows context.

Functions

build_flow_objects(json, uuid_map, process_fn, object \\ nil)

@spec build_flow_objects(
  list(),
  map(),
  (map(), map(), any() -> {any(), map()}),
  any()
) ::
  {any(), map()}

A generic json traversal and building the structure for a specific flow schema which is an array of objects in the json file. Used for Node/Actions, Node/Exits, Router/Cases, and Router/Categories

change_flow(flow, attrs \\ %{})

@spec change_flow(Glific.Flows.Flow.t(), map()) :: Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking flow changes.

Examples

iex> change_flow(flow)
%Ecto.Changeset{data: %Flow{}}

check_required_fields(json, required)

@spec check_required_fields(map(), [atom()]) :: boolean()

Check the required fields for all flow objects. If missing, raise an exception

copy_flow(flow, attrs)

@spec copy_flow(Glific.Flows.Flow.t(), map()) ::
  {:ok, Glific.Flows.Flow.t()} | {:error, String.t()}

Make a copy of a flow

count_flows(args)

@spec count_flows(map()) :: integer()

Return the count of flows, using the same filter as list_flows

create_flow(attrs)

@spec create_flow(map()) ::
  {:ok, Glific.Flows.Flow.t()} | {:error, Ecto.Changeset.t()}

Creates a flow.

Examples

iex> create_flow(%{field: value})
{:ok, %Flow{}}

iex> create_flow(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_flow_revision(definition)

@spec create_flow_revision(map()) :: Glific.Flows.FlowRevision.t()

Save new revision for the flow

delete_flow(flow)

@spec delete_flow(Glific.Flows.Flow.t()) ::
  {:ok, Glific.Flows.Flow.t()} | {:error, Ecto.Changeset.t()}

Deletes a flow.

Examples

iex> delete_flow(flow)
{:ok, %Flow{}}

iex> delete_flow(flow)
{:error, %Ecto.Changeset{}}

export_flow(flow_id)

@spec export_flow(non_neg_integer()) :: map()

Generate a json map with all the flows related fields.

export_flow_details(flow_uuid, results)

@spec export_flow_details(String.t(), map()) :: map()

process subflows and check if there is more subflows in it.

fetch_flow(id)

@spec fetch_flow(integer()) :: {:ok, any()} | {:error, any()}

Fetches a single flow

Returns Resource not found if the Interactive Template does not exist.

Examples

iex> fetch_interactive_template(123, 1)
  {:ok, %Flow{}}

iex> fetch_interactive_template(456, 1)
  {:error, ["Elixir.Glific.Flows.Flow", "Resource not found"]}

flow_activated(flow_id, contact_id, since)

@spec flow_activated(non_neg_integer(), non_neg_integer(), DateTime.t()) :: boolean()

Check if a flow has been activated since the time sent as a parameter e.g. outofoffice will check if that flow was activated in the last 24 hours daily/weekly will check since start of day/week, etc

flow_keywords_map(organization_id)

@spec flow_keywords_map(non_neg_integer()) :: map()

Create a map of keywords that map to flow ids for each active organization. Also cache this value including the outoffice shortcode

get_cached_flow(organization_id, key)

@spec get_cached_flow(non_neg_integer(), {atom(), any(), String.t()}) ::
  {atom(), any()} | {atom(), String.t()}

A helper function to interact with the Caching API and get the cached flow. It will also set the loaded flow to cache in case it does not exists.

get_flow!(id)

@spec get_flow!(integer()) :: Glific.Flows.Flow.t()

Gets a single flow.

Raises Ecto.NoResultsError if the Flow does not exist.

Examples

iex> get_flow!(123)
%Flow{}

iex> get_flow!(456)
** (Ecto.NoResultsError)

get_flow_revision(flow_uuid, revision_id)

@spec get_flow_revision(String.t(), String.t()) :: map()

Get specific flow revision by number

get_flow_revision_list(flow_uuid)

@spec get_flow_revision_list(String.t()) :: %{results: list()}

Get a list of all the revisions based on a flow UUID

get_sub_flows(nodes)

@spec get_sub_flows(list()) :: list()

Extract all the subflows form the parent flow definition.

import_flow(import_flow, organization_id)

@spec import_flow(map(), non_neg_integer()) :: boolean()

import a flow from json

init_export_flow(results, flow_uuid)

@spec init_export_flow(map(), String.t()) :: map()

Process the flows and get all the subflow definition.

is_media_type?(type)

@spec is_media_type?(atom()) :: boolean()

Check if the type is a media type we handle in flows

is_optin_flow?(flow)

@spec is_optin_flow?(Glific.Flows.Flow.t()) :: boolean()

Check if the flow is optin flow. Currently we are checking based on the optin keyword only.

list_flows(args)

@spec list_flows(map()) :: [Glific.Flows.Flow.t()]

Returns the list of flows.

Examples

iex> list_flows()
[%Flow{}, ...]

publish_flow(flow)

@spec publish_flow(Glific.Flows.Flow.t()) ::
  {:ok, Glific.Flows.Flow.t()} | {:error, any()}

Update latest flow revision status as published and increment the version Update cached flow definition

start_contact_flow(flow_id, contact)

@spec start_contact_flow(
  Glific.Flows.Flow.t() | integer(),
  Glific.Contacts.Contact.t()
) ::
  {:ok, Glific.Flows.Flow.t()} | {:error, String.t()}

Start flow for a contact

start_group_flow(flow, group)

@spec start_group_flow(Glific.Flows.Flow.t(), Glific.Groups.Group.t()) ::
  {:ok, Glific.Flows.Flow.t()}

Start flow for contacts of a group

terminate_contact_flows?(contact_id)

@spec terminate_contact_flows?(non_neg_integer()) :: :ok

Terminate all flows for a contact

update_cached_flow(flow, status)

@spec update_cached_flow(Glific.Flows.Flow.t(), String.t()) :: {atom(), any()}

Update the cached flow from db. This typically happens when the flow definition is updated via the UI

update_flow(flow, attrs)

@spec update_flow(Glific.Flows.Flow.t(), map()) ::
  {:ok, Glific.Flows.Flow.t()} | {:error, Ecto.Changeset.t()}

Updates a flow.

Examples

iex> update_flow(flow, %{field: new_value})
{:ok, %Flow{}}

iex> update_flow(flow, %{field: bad_value})
{:error, %Ecto.Changeset{}}