Glific.Flows.Broadcast (Glific v4.3.0)

Start a flow to a group so we can blast it out as soon as possible and ensure we are under the rate limits.

Functions

broadcast_contacts(flow, contacts, opts \\ [])

@spec broadcast_contacts(map(), [Glific.Contacts.Contact.t()], Keyword.t()) :: :ok

Lets start a bunch of contacts on a flow in parallel

broadcast_group(flow, group)

@spec broadcast_group(Glific.Flows.Flow.t(), Glific.Groups.Group.t()) :: map()

The one simple public interface to broadcast a group

execute_group_broadcasts(org_id)

@spec execute_group_broadcasts(any()) :: :ok

The one simple public interface to exceute a group broadcast for an organization

mark_flow_broadcast_completed(org_id)

@spec mark_flow_broadcast_completed(non_neg_integer()) :: :ok

Mark all the proceesed flow broadcast as completed

process_broadcast_group(flow_broadcast)

@spec process_broadcast_group(Glific.Flows.FlowBroadcast.t() | nil) :: :ok

Start a group broadcast for a giving broadcast stuct