Glific.Flows.Case (Glific v0.9.4)

The Case object which encapsulates one category in a given node.

Types

t()

Specs

t() :: %Glific.Flows.Case{
  arguments: [String.t()],
  category: Glific.Flows.Category.t() | nil,
  category_uuid: Ecto.UUID.t() | nil,
  id: term(),
  name: term(),
  type: Glific.Enums.FlowCase | nil,
  uuid: Ecto.UUID.t() | nil
}

Functions

execute(c, context, msg)

Specs

execute(t(), Glific.Flows.FlowContext.t(), Glific.Messages.Message.t()) ::
  boolean()

Execute a case, given a message. This is the only execute function which has a different signature, since it just consumes one message at a time and executes it against a predefined function It also returns a boolean, rather than a tuple

process(json, uuid_map, object \\ nil)

Specs

process(map(), map(), any()) :: {t(), map()}

Process a json structure from floweditor to the Glific data types