Glific v0.2.0 Glific.Flows.Category

The Category object which encapsulates one category in a given node.

Types

t()

Specs

t() :: %Glific.Flows.Category{
  cases: [Glific.Flows.Case.t()] | [],
  exit: Glific.Flows.Exit.t() | nil,
  exit_uuid: Ecto.UUID.t() | nil,
  id: term(),
  name: String.t() | nil,
  uuid: Ecto.UUID.t() | nil
}

Functions

execute(category, context, message_stream)

Specs

execute(t(), Glific.Flows.FlowContext.t(), [String.t()]) ::
  {:ok, Glific.Flows.FlowContext.t(), [String.t()]} | {:error, String.t()}

Execute a category, given a message stream.

process(json, uuid_map, object \\ nil)

Specs

process(map(), map(), any()) :: {t(), map()}

Process a json structure from floweditor to the Glific data types