Glific.Flows.ContactAction (Glific v0.9.4)

Since many of the functions, also do a few actions like send a message etc centralizing it here

Functions

optout(context)

Specs

optout(Glific.Flows.FlowContext.t()) :: Glific.Flows.FlowContext.t()

Contact opts out

send_broadcast(context, action, messages)

Specs

send_broadcast(Glific.Flows.FlowContext.t(), Glific.Flows.Action.t(), [
  Glific.Messages.Message.t()
]) :: {:ok, map(), any()}

This is just a think wrapper for send_message, since its basically the same, but instead of sending the message to the contact, sends it to another contact that is identified in the action. You can send the same notification to multiple contacts

send_message(context, action, messages, cid \\ nil)

Specs

send_message(
  Glific.Flows.FlowContext.t(),
  Glific.Flows.Action.t(),
  [Glific.Messages.Message.t()],
  non_neg_integer() | nil
) :: {:ok, map(), any()}

If the template is not defined for the message send text messages. Given a shortcode and a context, send the right session template message to the contact.

We also need to handle translations for template messages, since whatsapp gives them unique uuids