Glific.Flows.ContactField (Glific v4.3.0)

Since many of the functions set/update fields in contact and related tables, lets centralize all the code here for now

Functions

add_contact_field(context, field, label, value, type)

@spec add_contact_field(
  Glific.Flows.FlowContext.t(),
  String.t(),
  String.t(),
  String.t(),
  String.t()
) ::
  Glific.Flows.FlowContext.t()

Add a field {key, value} to a contact. For now, all preferences are stored under the settings map, with a sub-map of preferences. We expect to get more clarity on this soon

count_contacts_fields(args)

@spec count_contacts_fields(map()) :: integer()

Return the count of contacts_fields, using the same filter as list_contacts_fields

create_contact_field(attrs)

@spec create_contact_field(map()) ::
  {:ok, Glific.Contacts.ContactsField.t()} | {:error, Ecto.Changeset.t()}

Create contact field

delete_contacts_field(contacts_field)

@spec delete_contacts_field(Glific.Contacts.ContactsField.t()) ::
  {:ok, Glific.Contacts.ContactsField.t()} | {:error, Ecto.Changeset.t()}

Deletes a contact field.

Examples

iex> delete_contacts_field(contacts_field)
{:ok, %ContactsField{}}

iex> delete_contacts_field(contacts_field)
{:error, %Ecto.Changeset{}}

do_add_contact_field(contact, field, label, value, type \\ "string")

@spec do_add_contact_field(
  Glific.Contacts.Contact.t(),
  String.t(),
  String.t(),
  any(),
  String.t()
) ::
  Glific.Contacts.Contact.t()

Add contact field taking contact as parameter. We should change the name of this function for the consistency

list_contacts_fields(args)

@spec list_contacts_fields(map()) :: [Glific.Contacts.ContactsField.t()]

list contacts fields.

parse_contact_field_value(context, value)

@spec parse_contact_field_value(Glific.Flows.FlowContext.t(), String.t()) ::
  String.t()

parse contact fields values with check if it has

reset_contact_fields(context)

@spec reset_contact_fields(Glific.Flows.FlowContext.t()) ::
  Glific.Flows.FlowContext.t()

Reset the fields for a contact.

update_contacts_field(contacts_field, attrs)

@spec update_contacts_field(Glific.Contacts.ContactsField.t(), map()) ::
  {:ok, Glific.Contacts.ContactsField.t()} | {:error, Ecto.Changeset.t()}

Updates a contact field.

Examples

iex> update_contacts_field(contacts_field, %{field: new_value})
{:ok, %ContactsField{}}

iex> update_contacts_field(contacts_field, %{field: bad_value})
{:error, %Ecto.Changeset{}}