Glific.Flows.Flow (Glific v4.3.0)

The flow object which encapsulates the complete flow as emitted by by https://github.com/nyaruka/floweditor

Types

t()

@type t() :: %Glific.Flows.Flow{
  __meta__: Ecto.Schema.Metadata.t(),
  definition: map() | nil,
  flow_type: String.t() | nil,
  id: non_neg_integer() | nil,
  ignore_keywords: boolean() | nil,
  inserted_at: :utc_datetime | nil,
  is_active: boolean() | nil,
  is_background: boolean() | nil,
  keywords: [String.t()] | nil,
  last_changed_at: term(),
  last_published_at: term(),
  localization: Glific.Flows.Localization.t() | nil,
  name: String.t() | nil,
  nodes: [Glific.Flows.Node.t()] | nil,
  organization:
    Glific.Partners.Organization.t() | Ecto.Association.NotLoaded.t() | nil,
  organization_id: non_neg_integer() | nil,
  respond_no_response: boolean() | nil,
  respond_other: boolean() | nil,
  revisions:
    [Glific.Flows.FlowRevision.t()] | Ecto.Association.NotLoaded.t() | nil,
  start_node: Glific.Flows.Node.t() | nil,
  status: String.t(),
  updated_at: :utc_datetime | nil,
  uuid: Ecto.UUID.t() | nil,
  uuid_map: map() | nil,
  version: term(),
  version_number: String.t() | nil
}

Functions

changeset(flow, attrs)

@spec changeset(t(), map()) :: Ecto.Changeset.t()

Standard changeset pattern we use for all data types

get_flow(organization_id, uuid, status)

@spec get_flow(non_neg_integer(), Ecto.UUID.t(), String.t()) :: map()

Return a flow for a specific uuid. Cache is not present in cache

get_latest_definition(flow_id)

@spec get_latest_definition(integer()) :: map()

load the latest revision, specifically json definition from the flow_revision table. We return the clean definition back

get_loaded_flow(organization_id, status, args)

@spec get_loaded_flow(non_neg_integer(), String.t(), map()) :: map()

Helper function to load a active flow from the database and build an object

process(json, flow, start_node_uuid)

@spec process(map(), t(), Ecto.UUID.t()) :: t()

Process a json structure from floweditor to the Glific data types. While we are doing this we also fix the map, if the variables to resolve Other/No Response is true

start_sub_flow(context, uuid, parent_id)

@spec start_sub_flow(Glific.Flows.FlowContext.t(), Ecto.UUID.t(), non_neg_integer()) ::
  {:ok, Glific.Flows.FlowContext.t(), [String.t()]} | {:error, String.t()}

Create a subflow of an existing flow

validate_flow(organization_id, status, args)

@spec validate_flow(non_neg_integer(), String.t(), map()) :: Keyword.t()

Validate a flow and ensures the flow is valid with our internal rule-set

validate_keywords(changeset, keywords)

@spec validate_keywords(Ecto.Changeset.t(), any()) :: Ecto.Changeset.t()

Changeset helper for keywords