Glific.Flows.FlowLabel (Glific v0.9.4)

The flow label object

Types

t()

Specs

t() :: %Glific.Flows.FlowLabel{
  __meta__: Ecto.Schema.Metadata.t(),
  id: non_neg_integer() | nil,
  inserted_at: :utc_datetime | nil,
  name: String.t() | nil,
  organization:
    Glific.Partners.Organization.t() | Ecto.Association.NotLoaded.t() | nil,
  organization_id: non_neg_integer() | nil,
  updated_at: :utc_datetime | nil,
  uuid: Ecto.UUID.t() | nil
}

Functions

changeset(flow_label, attrs)

Specs

changeset(any(), map()) :: Ecto.Changeset.t()

Glific.Flows.FlowLabel.list_flow_label() Standard changeset pattern we use for all data types

create_flow_label(attrs)

Specs

create_flow_label(map()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Creates a flow_label.

Examples

iex> create_flow_label(%{field: value})
{:ok, %FlowLabel{}}

iex> create_flow_label(%{field: bad_value})
{:error, %Ecto.Changeset{}}

get_all_flowlabel(organization_id)

Specs

get_all_flowlabel(non_neg_integer()) :: [t()]

Given a organization id, retrieve all the flow labels for organization

get_or_create_flow_label(attrs)

Specs

get_or_create_flow_label(map()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Try to first get the flow label, if not present, create it. We dont use the upsert function, since it consumes id's for every failure. we expect a lot more gets, than inserts