Glific v0.2.0 Glific.Flows.Node

The Node object which encapsulates one node in a given flow

Types

t()

Specs

t() :: %Glific.Flows.Node{
  actions: [Glific.Flows.Action.t()] | [],
  exits: [Glific.Flows.Exit.t()] | [],
  flow: Glific.Flows.Flow.t() | nil,
  flow_uuid: Ecto.UUID.t() | nil,
  id: term(),
  router: Glific.Flows.Router.t() | nil,
  uuid: Ecto.UUID.t() | nil
}

Functions

execute(node, context, message_stream)

Specs

execute(t(), Glific.Flows.FlowContext.t(), [String.t()]) ::
  {:ok | :wait, Glific.Flows.FlowContext.t(), [String.t()]}
  | {:error, String.t()}

Execute a node, given a message stream. Consume the message stream as processing occurs

process(json, uuid_map, flow)

Specs

process(map(), map(), Glific.Flows.Flow.t()) :: {t(), map()}

Process a json structure from floweditor to the Glific data types