Glific.Flows.Router (Glific v0.9.4)

The Router object which encapsulates the router in a given node.

Types

t()

Specs

t() :: %Glific.Flows.Router{
  __meta__: term(),
  cases: [Glific.Flows.Case.t()] | nil,
  categories: [Glific.Flows.Category.t()] | nil,
  default_category: Glific.Flows.Category.t() | nil,
  default_category_uuid: Ecto.UUID.t() | nil,
  id: term(),
  node: Glific.Flows.Node.t() | nil,
  node_uuid: Ecto.UUID.t() | nil,
  operand: term(),
  result_name: String.t() | nil,
  type: String.t() | nil,
  wait: Glific.Flows.Wait.t() | nil,
  wait_type: term()
}

Functions

execute(router, context, messages)

Specs

execute(t(), Glific.Flows.FlowContext.t(), [Glific.Messages.Message.t()]) ::
  {:ok, Glific.Flows.FlowContext.t(), [Glific.Messages.Message.t()]}
  | {:error, String.t()}

Execute a router, given a message stream. Consume the message stream as processing occurs

process(json, uuid_map, node)

Specs

process(map(), map(), Glific.Flows.Node.t()) :: {t(), map()}

Process a json structure from floweditor to the Glific data types