Glific.Groups (Glific v4.3.0)

The Groups context.

Functions

add_permission(query, user)

@spec add_permission(Ecto.Query.t(), Glific.Users.User.t()) :: Ecto.Query.t()

Add permissioning specific to groups, in this case we want to restrict the visibility of groups that the user can see

change_group(group, attrs \\ %{})

@spec change_group(Glific.Groups.Group.t(), map()) :: Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking group changes.

Examples

iex> change_group(group)
%Ecto.Changeset{data: %Group{}}

contact_ids(group_id)

@spec contact_ids(non_neg_integer()) :: [non_neg_integer()]

Get the contacts ids for a specific group that have not opted out

contacts_count(map)

@spec contacts_count(map()) :: integer()

Return the count of group contacts

count_groups(args)

@spec count_groups(map()) :: integer()

Return the count of groups, using the same filter as list_groups

create_contact_group(attrs \\ %{})

@spec create_contact_group(map()) ::
  {:ok, Glific.Groups.ContactGroup.t()} | {:error, Ecto.Changeset.t()}

Creates a contact group.

Examples

iex> create_contact_group(%{field: value})
{:ok, %Contact{}}

iex> create_contact_group(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_group(attrs \\ %{})

@spec create_group(map()) ::
  {:ok, Glific.Groups.Group.t()} | {:error, Ecto.Changeset.t()}

Creates a group.

Examples

iex> create_group(%{field: value})
{:ok, %Group{}}

iex> create_group(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_user_group(attrs \\ %{})

@spec create_user_group(map()) ::
  {:ok, Glific.Groups.UserGroup.t()} | {:error, Ecto.Changeset.t()}

Creates a user group.

Examples

iex> create_user_group(%{field: value})
{:ok, %User{}}

iex> create_user_group(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_contact_groups_by_ids(contact_id, group_ids)

@spec delete_contact_groups_by_ids(integer(), list()) :: {integer(), nil | [term()]}

Delete contact groups

delete_group(group)

@spec delete_group(Glific.Groups.Group.t()) ::
  {:ok, Glific.Groups.Group.t()} | {:error, Ecto.Changeset.t()}

Deletes a group.

Examples

iex> delete_group(group)
{:ok, %Group{}}

iex> delete_group(group)
{:error, %Ecto.Changeset{}}

delete_group_contacts_by_ids(group_id, contact_ids)

@spec delete_group_contacts_by_ids(integer(), list()) :: {integer(), nil | [term()]}

Delete group contacts

delete_group_users_by_ids(group_id, user_ids)

@spec delete_group_users_by_ids(integer(), []) :: {integer(), nil | [term()]}

Delete group users

delete_user_groups_by_ids(user_id, group_ids)

@spec delete_user_groups_by_ids(integer(), []) :: {integer(), nil | [term()]}

Delete user groups

get_group!(id)

@spec get_group!(integer()) :: Glific.Groups.Group.t()

Gets a single group.

Raises Ecto.NoResultsError if the Group does not exist.

Examples

iex> get_group!(123)
%Group{}

iex> get_group!(456)
** (Ecto.NoResultsError)

get_group_ids()

@spec get_group_ids() :: list()

Fetches all group ids in an organization

get_or_create_group_by_label(label, organization_id)

@spec get_or_create_group_by_label(String.t(), non_neg_integer()) ::
  {:ok, Glific.Groups.Group.t()} | nil

Get group by group name. Create the group if it does not exist

info_group_contacts(group_id)

@spec info_group_contacts(non_neg_integer()) :: map()

Given a group id, get stats on the contacts within this group based on bsp_status and also the total count

list_groups(args, skip_permission \\ false)

@spec list_groups(map(), boolean()) :: [Glific.Groups.Group.t()]

Returns the list of groups.

Examples

iex> list_groups()
[%Group{}, ...]

list_organizations_groups(args)

@spec list_organizations_groups(map()) :: [Glific.Groups.Group.t()]

Returns the list of groups.

Examples

iex> list_organizations_groups()
[%Group{}, ...]

update_group(group, attrs)

@spec update_group(Glific.Groups.Group.t(), map()) ::
  {:ok, Glific.Groups.Group.t()} | {:error, Ecto.Changeset.t()}

Updates a group.

Examples

iex> update_group(group, %{field: new_value})
{:ok, %Group{}}

iex> update_group(group, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_user_groups(map)

@spec update_user_groups(map()) :: :ok

Updates user groups entries

users_count(map)

@spec users_count(map()) :: integer()

Return the count of group users