Glific.Jobs.MinuteWorker (Glific v4.3.0)

Processes the tasks that need to be handled on a minute schedule

Functions

perform(args)

@spec perform(Oban.Job.t()) ::
  :discard | :ok | {:error, any()} | {:ok, any()} | {:snooze, pos_integer()}

Worker to implement cron job functionality as implemented by Oban. This is a work in progress and subject to change