Glific v0.2.0 Glific.Partners.Organization

Organizations are the group of users who will access the system

Types

t()

Specs

t() :: %Glific.Partners.Organization{
  __meta__: Ecto.Schema.Metadata.t(),
  contact: Glific.Contacts.Contact.t() | Ecto.Association.NotLoaded.t() | nil,
  contact_id: non_neg_integer() | nil,
  contact_name: String.t() | nil,
  default_language:
    Glific.Settings.Language.t() | Ecto.Association.NotLoaded.t() | nil,
  default_language_id: non_neg_integer() | nil,
  display_name: String.t() | nil,
  email: String.t() | nil,
  id: non_neg_integer() | nil,
  inserted_at: :utc_datetime | nil,
  name: String.t() | nil,
  provider: Glific.Partners.Provider.t() | Ecto.Association.NotLoaded.t() | nil,
  provider_id: non_neg_integer() | nil,
  provider_key: String.t() | nil,
  provider_number: String.t() | nil,
  updated_at: :utc_datetime | nil
}

Functions

changeset(organization, attrs)

Specs

changeset(t(), map()) :: Ecto.Changeset.t()

Standard changeset pattern we use for all datat types