Glific v0.1.0 Glific.Processor.Helper

Helper functions for all processing modules. Might promote this up at a later stage

Functions

add_tag(message, tag_id, value \\ nil)

Specs

add_tag(Glific.Messages.Message.t(), integer(), String.t() | nil) ::
  Glific.Messages.Message.t()

Helper function to add tag

get_session_message_template(shortcode, language_id \\ nil)

Specs

get_session_message_template(String.t(), integer() | nil) ::
  Glific.Templates.SessionTemplate.t()

Given a shortcode and an optional language_id, get the session template matching both, and if not found, just for the shortcode

send_session_message_template(message, shortcode)

Specs

send_session_message_template(Glific.Messages.Message.t(), String.t()) ::
  Glific.Messages.Message.t()

Send a reply to the current sender of the incoming message in the preferred language of the sender

send_session_message_template_with_tag(message, tag, value, shortcode)

Specs

send_session_message_template_with_tag(
  Glific.Messages.Message.t(),
  Glific.Tags.Tag.t(),
  String.t(),
  String.t()
) :: Glific.Messages.Message.t()

Send a reply to the current sender of the incoming message in the preferred language of the sender and associate a tag with it

start_link(opts, module)

Specs

start_link([], atom()) :: GenServer.on_start()

Common function for downstream consumers since they follow the same pattern