Glific v0.1.0 Glific.Processor.Producer

This producer is linked to the message receiving end and gets messages from the external world in the message structure

Functions

add(messages)

Specs

add([Glific.Messages.Message.t()]) :: :ok

add(Glific.Messages.Message.t()) :: :ok

public endpoint for adding a new message or a set of messages

handle_cast(arg, state)

Specs

handle_cast({:add, [Glific.Messages.Message.t()]}, map()) ::
  {:noreply, [Glific.Messages.Message.t()], map()}

push a message to all consumers on adding

handle_demand(_, state)

Specs

handle_demand(integer(), any()) :: {:noreply, [], map()}

ignore all requests from consumers via demand call