Glific.Profiles (Glific v4.6.8)

The Profiles context.

Functions

create_profile(attrs \\ %{})

@spec create_profile(map()) ::
  {:ok, Glific.Profiles.Profile.t()} | {:error, Ecto.Changeset.t()}

Creates a profile.

Examples

iex> create_profile(%{field: value})
{:ok, %Profile{}}

iex> create_profile(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_profile(profile)

@spec delete_profile(Glific.Profiles.Profile.t()) ::
  {:ok, Glific.Profiles.Profile.t()} | {:error, Ecto.Changeset.t()}

Deletes a profile.

Examples

iex> delete_profile(profile)
{:ok, %Profile{}}

iex> delete_profile(profile)
{:error, %Ecto.Changeset{}}

get_indexed_profile(contact)

@spec get_indexed_profile(Glific.Contacts.Contact.t()) :: [{any(), integer()}]

Get a profile associated with a contact indexed and sorted in ascending order

Examples

iex> Glific.Profiles.get_indexed_profile(con)
[{%Profile{}, 1}, {%Profile{}, 2}]

get_profile!(id)

@spec get_profile!(integer()) :: Glific.Profiles.Profile.t()

Gets a single profile.

Raises Ecto.NoResultsError if the Profile does not exist.

Examples

iex> get_profile!(123)
%Profile{}

iex> get_profile!(456)
** (Ecto.NoResultsError)

handle_flow_action(arg1, context, action)

@spec handle_flow_action(
  atom() | nil,
  Glific.Flows.FlowContext.t(),
  Glific.Flows.Action.t()
) ::
  {Glific.Flows.FlowContext.t(), Glific.Messages.Message.t()}

Handles flow action based on type of operation on Profile

list_profiles(args)

@spec list_profiles(map()) :: [Glific.Profiles.Profile.t()]

Returns the list of profiles.

Examples

iex> list_profiles()
[%Profile{}, ...]

Get the list of profiles filtered by various search options

switch_profile(contact, profile_index)

@spec switch_profile(Glific.Contacts.Contact.t(), String.t()) ::
  Glific.Contacts.Contact.t()

Switches active profile of a contact

Examples

iex> switch_profile(contact)
{:ok, %Profile{}}

update_profile(profile, attrs)

@spec update_profile(Glific.Profiles.Profile.t(), map()) ::
  {:ok, Glific.Profiles.Profile.t()} | {:error, Ecto.Changeset.t()}

Updates a profile.

Examples

iex> update_profile(profile, %{field: new_value})
{:ok, %Profile{}}

iex> update_profile(profile, %{field: bad_value})
{:error, %Ecto.Changeset{}}