Glific v0.2.0 Glific.Providers.Glifproxy.Worker

A worker to handle send message processes

Functions

perform(job)

Specs

perform(Oban.Job.t()) :: :ok

Standard perform method to use Oban worker

proxy_message(message, payload)

Specs

proxy_message(map(), Oban.Job.t()) :: any()

We transform a payload that has been set for sending to a payload that has been tailored for receiving. We need to generate a few random ids for various messages ids