Glific.Repo (Glific v0.9.4)

A repository that maps to an underlying data store, controlled by the Postgres adapter.

We add a few functions to make our life easier with a few helper functions that ecto does not provide.

Functions

add_permission(query, permission_fn, skip_permission \\ false)

Specs

add_permission(
  Ecto.Query.t(),
  (Ecto.Query.t(), Glific.Users.User.t() -> Ecto.Query.t()),
  boolean()
) :: Ecto.Query.t()

Implement permissioning support via groups. This is the basic wrapper, it uses a context specific permissioning wrapper to add the actual clauses

aggregate(queryable, aggregate, opts \\ [])

Callback implementation for Ecto.Repo.aggregate/3.

aggregate(queryable, aggregate, field, opts)

Callback implementation for Ecto.Repo.aggregate/4.

all(queryable, opts \\ [])

Callback implementation for Ecto.Repo.all/2.

checkout(fun, opts \\ [])

Callback implementation for Ecto.Repo.checkout/2.

child_spec(opts)

config()

Callback implementation for Ecto.Repo.config/0.

count_filter(args \\ %{}, object, filter_with_fn, repo_opts \\ [])

Specs

count_filter(
  map(),
  atom(),
  (Ecto.Queryable.t(), %{optional(atom()) => any()} -> Ecto.Queryable.t()),
  Keyword.t()
) :: integer()

We use this function also in most list_OBJECT api's, where we process the the filter. Centralizing this code at the top level, to make things cleaner

delete(struct, opts \\ [])

Callback implementation for Ecto.Repo.delete/2.

delete!(struct, opts \\ [])

Callback implementation for Ecto.Repo.delete!/2.

delete_all(queryable, opts \\ [])

Callback implementation for Ecto.Repo.delete_all/2.

delete_relationships_by_ids(object, fields)

Specs

delete_relationships_by_ids(
  atom(),
  {{atom(), integer()}, {atom(), [integer()]}}
) :: {integer(), nil | [term()]}

In Join tables we rarely use the table id. We always know the object ids and hence more convenient to delete an entry via its object ids.

exists?(queryable, opts \\ [])

Callback implementation for Ecto.Repo.exists?/2.

explain(operation, queryable, opts \\ [])

A convenience function for SQL-based repositories that executes an EXPLAIN statement or similar depending on the adapter to obtain statistics for the given query.

See Ecto.Adapters.SQL.explain/4 for more information.

fetch(queryable, id, opts \\ [])

Specs

fetch(Ecto.Queryable.t(), term(), Keyword.t()) ::
  {:ok, Ecto.Schema.t()} | {:error, [String.t()]}

Glific version of get, which returns a tuple with an :ok | :error as the first element

fetch_by(queryable, clauses, opts \\ [])

Specs

fetch_by(Ecto.Queryable.t(), Keyword.t() | map(), Keyword.t()) ::
  {:ok, Ecto.Schema.t()} | {:error, [String.t()]}

Glific version of get_by, which returns a tuple with an :ok | :error as the first element

filter_with(query, filter)

Specs

filter_with(Ecto.Queryable.t(), %{optional(atom()) => any()}) ::
  Ecto.Queryable.t()

Add all the common filters here, rather than in each file

get(queryable, id, opts \\ [])

Callback implementation for Ecto.Repo.get/3.

get!(queryable, id, opts \\ [])

Callback implementation for Ecto.Repo.get!/3.

get_by(queryable, clauses, opts \\ [])

Callback implementation for Ecto.Repo.get_by/3.

get_by!(queryable, clauses, opts \\ [])

Callback implementation for Ecto.Repo.get_by!/3.

get_dynamic_repo()

Callback implementation for Ecto.Repo.get_dynamic_repo/0.

in_transaction?()

Callback implementation for Ecto.Repo.in_transaction?/0.

insert(struct, opts \\ [])

Callback implementation for Ecto.Repo.insert/2.

insert!(struct, opts \\ [])

Callback implementation for Ecto.Repo.insert!/2.

insert_all(schema_or_source, entries, opts \\ [])

Callback implementation for Ecto.Repo.insert_all/3.

insert_or_update(changeset, opts \\ [])

Callback implementation for Ecto.Repo.insert_or_update/2.

insert_or_update!(changeset, opts \\ [])

Callback implementation for Ecto.Repo.insert_or_update!/2.

label_id_map(queryable, values, organization_id, field \\ :label)

Specs

label_id_map(Ecto.Queryable.t(), [String.t()], non_neg_integer(), atom()) :: %{
  required(String.t()) => integer()
}

Get map of field (typically label) to ids for easier lookup for various system objects - language, tag

limit_offset(query, opts)

Specs

limit_offset(Ecto.Queryable.t(), map()) :: Ecto.Queryable.t()

Extracts the limit offset field, and adds to query

list_filter(args \\ %{}, object, opts_with_fn, filter_with_fn, repo_opts \\ [])

Specs

list_filter(
  map(),
  atom(),
  (Ecto.Queryable.t(), %{optional(atom()) => any()} -> Ecto.Queryable.t()),
  (Ecto.Queryable.t(), %{optional(atom()) => any()} -> Ecto.Queryable.t()),
  Keyword.t()
) :: [any()]

We use this function in most list_OBJECT api's, where we process the opts and the filter. Centralizing this code at the top level, to make things cleaner

list_filter_query(args \\ %{}, object, opts_with_fn, filter_with_fn)

Specs

list_filter_query(
  map(),
  atom(),
  (Ecto.Queryable.t(), %{optional(atom()) => any()} -> Ecto.Queryable.t())
  | nil,
  (Ecto.Queryable.t(), %{optional(atom()) => any()} -> Ecto.Queryable.t())
) :: Ecto.Queryable.t()

This function builds the query, and is used in places where we want to layer permissioning on top of the query

load(schema_or_types, data)

Callback implementation for Ecto.Repo.load/2.

one(queryable, opts \\ [])

Callback implementation for Ecto.Repo.one/2.

one!(queryable, opts \\ [])

Callback implementation for Ecto.Repo.one!/2.

opts_with_field(query, opts, field)

Specs

opts_with_field(
  Ecto.Queryable.t(),
  map(),
  :name | :body | :label | :inserted_at
) :: Ecto.Queryable.t()

A funtion which handles the order clause for a data type that has a 'name/body/label' in its schema (which is true for a fair number of Glific's data types)

opts_with_nil(query, opts)

Specs

opts_with_nil(Ecto.Queryable.t(), any()) :: Ecto.Queryable.t()

An empty function for objects that ignore the opts

preload(struct_or_structs_or_nil, preloads, opts \\ [])

Callback implementation for Ecto.Repo.preload/3.

put_dynamic_repo(dynamic)

Callback implementation for Ecto.Repo.put_dynamic_repo/1.

query(sql, params \\ [], opts \\ [])

A convenience function for SQL-based repositories that executes the given query.

See Ecto.Adapters.SQL.query/4 for more information.

query!(sql, params \\ [], opts \\ [])

A convenience function for SQL-based repositories that executes the given query.

See Ecto.Adapters.SQL.query!/4 for more information.

reload(queryable, opts \\ [])

Callback implementation for Ecto.Repo.reload/2.

reload!(queryable, opts \\ [])

Callback implementation for Ecto.Repo.reload!/2.

rollback(value)

Specs

rollback(term()) :: no_return()

Callback implementation for Ecto.Repo.rollback/1.

skip_permission?()

Specs

skip_permission?() :: Glific.Users.User.t() | true

Can we skip checking permissions for this user. This eliminates a DB call in many a case

start_link(opts \\ [])

Callback implementation for Ecto.Repo.start_link/1.

stop(timeout \\ 5000)

Callback implementation for Ecto.Repo.stop/1.

stream(queryable, opts \\ [])

Callback implementation for Ecto.Repo.stream/2.

to_sql(operation, queryable)

A convenience function for SQL-based repositories that translates the given query to SQL.

See Ecto.Adapters.SQL.to_sql/3 for more information.

transaction(fun_or_multi, opts \\ [])

Callback implementation for Ecto.Repo.transaction/2.

update(struct, opts \\ [])

Callback implementation for Ecto.Repo.update/2.

update!(struct, opts \\ [])

Callback implementation for Ecto.Repo.update!/2.

update_all(queryable, updates, opts \\ [])

Callback implementation for Ecto.Repo.update_all/3.