Glific.Searches (Glific v0.9.4)

The Searches context.

Functions

add_permission(query, user)

Specs

add_permission(Ecto.Query.t(), Glific.Users.User.t()) :: Ecto.Query.t()

Add permissioning specific to searches, in this case we want to restrict the visibility of contact ids

change_saved_search(search, attrs \\ %{})

Specs

change_saved_search(Glific.Searches.SavedSearch.t(), map()) ::
  Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking search changes.

Examples

iex> change_saved_search(search)
%Ecto.Changeset{data: %Search{}}

convert_to_atom(json)

Specs

convert_to_atom(map()) :: map()

Given a jsonb string, typically either from the database, or maybe via graphql convert the string keys to atoms

count_saved_searches(args)

Specs

count_saved_searches(map()) :: integer()

Returns the count of searches, using the same filter as list_saved_searches

create_saved_search(attrs)

Specs

create_saved_search(map()) ::
  {:ok, Glific.Searches.SavedSearch.t()} | {:error, Ecto.Changeset.t()}

Creates a search.

Examples

iex> create_saved_search(%{field: value})
{:ok, %SavedSearch{}}

iex> create_saved_search(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_saved_search(search)

Specs

delete_saved_search(Glific.Searches.SavedSearch.t()) ::
  {:ok, Glific.Searches.SavedSearch.t()} | {:error, Ecto.Changeset.t()}

Deletes a search.

Examples

iex> delete_saved_search(search)
{:ok, %SavedSearch{}}

iex> delete_saved_search(search)
{:error, %Ecto.Changeset{}}

get_saved_search!(id)

Specs

get_saved_search!(integer()) :: Glific.Searches.SavedSearch.t()

Gets a single search.

Raises Ecto.NoResultsError if the SavedSearch does not exist.

Examples

iex> get_saved_search!(123)
%SavedSearch{}

iex> get_saved_search!(456)
** (Ecto.NoResultsError)

list_saved_searches(args)

Specs

list_saved_searches(map()) :: [Glific.Searches.SavedSearch.t()]

Returns the list of searches.

Examples

iex> list_saved_searches()
[%SavedSearch{}, ...]

saved_search_count(args)

Specs

saved_search_count(map()) :: [Glific.Conversations.Conversation.t()] | integer()

Execute a saved search, if term is sent in, it is added to the saved search. Either return conversations or count

search(args, count \\ false)

Specs

search(map(), boolean()) :: [Glific.Conversations.Conversation.t()] | integer()

Full text search interface via Postgres

search_multi(term, args)

Specs

search_multi(String.t(), map()) :: Glific.Searches.Search.t()

Search across multiple tables, and return a multiple context result back to the frontend. First step in emulating a whatsapp search

update_saved_search(search, attrs)

Specs

update_saved_search(Glific.Searches.SavedSearch.t(), map()) ::
  {:ok, Glific.Searches.SavedSearch.t()} | {:error, Ecto.Changeset.t()}

Updates a search.

Examples

iex> update_saved_search(search, %{field: new_value})
{:ok, %SavedSearch{}}

iex> update_saved_search(search, %{field: bad_value})
{:error, %Ecto.Changeset{}}