Glific.Seeds.SeedsMigration (Glific v4.3.0)

One shot migration of data to add simulators and saas admin. We use the functions in this file to add simulators for new organizations as they are created

Functions

add_saas_user(organization, language)

@spec add_saas_user(Glific.Partners.Organization.t(), Glific.Settings.Language.t()) ::
  :ok

Add a saas user for the organization. We need to check if it already exists since this code is used during data migration and can be repeated for the same organization

fix_message_number(org_id)

@spec fix_message_number(list() | integer()) :: :ok

Reset message number for a list of organizations or for a org_id

get_deleted_hsms(org_id)

@spec get_deleted_hsms(non_neg_integer()) :: tuple()

We need this functionality to cleanups all the Approved templates which are not active on Gupshup

migrate_data(phase, organization \\ nil)

@spec migrate_data(atom(), Glific.Partners.Organization.t() | nil) :: :ok

Public interface to do a seed migration across all organizations.

One function to rule them all. This function is invoked manually by a glific developer to add data from the DB. This seems the cleanest way to do such things. We use phases to seperate different migrations

sync_hsm_templates(org_id_list)

@spec sync_hsm_templates(list()) :: :ok

sync all the hsm from BSP to Glific DB

sync_schema_with_bigquery(org_id_list)

@spec sync_schema_with_bigquery(list()) :: :ok

Sync bigquery schema with local db changes.