Glific v0.1.0 Glific.SeedsDev

Script for populating the database. We can call this from tests and/or /priv/repo

Functions

seed()

Specs

seed() :: nil

Function to populate some basic data that we need for the system to operate. We will split this function up into multiple different ones for test, dev and production

seed_tag()

Specs

seed_tag() :: nil

Smaller functions to seed various tables. This allows the test functions to call specific seeder functions. In the next phase we will also add unseeder functions as we learn more of the test capabilities