Glific.Stats (Glific v1.3.3)

The stats manager and API to interface with the stat sub-system

Functions

count_stats(args)

Specs

count_stats(map()) :: integer()

Return the count of stats, using the same filter as list_stats

create_stat(attrs \\ %{})

Specs

create_stat(map()) ::
  {:ok, Glific.Stats.Stat.t()} | {:error, Ecto.Changeset.t()}

Create a Stat

generate_stats(list \\ [], recent \\ true, opts \\ [])

Specs

generate_stats(list(), boolean(), Keyword.t()) :: nil

Top level function to generate stats for all active organizations by default. Can control behavior by setting function parameters

list_stats(args)

Specs

list_stats(map()) :: list()

Returns the list of stats. Since this is very basic and only listing funcatinality we added the status filter like this. In future we will put the status as virtual filed in the stats itself.

update_stat(stat, attrs)

Specs

update_stat(Glific.Stats.Stat.t(), map()) ::
  {:ok, Glific.Stats.Stat.t()} | {:error, Ecto.Changeset.t()}

Update a Stat