Glific v0.1.0 Glific.Taggers

The API for a generic tagging system on messages that coordinate with different types of taggers. The proposed taggers are: Numeric Keyword Emojis

Positive
Negative

Automated

Compliments
Good Bye
Greeting
Thank You
Welcome
Spam

Functions

string_clean(str)

Specs

string_clean(String.t()) :: String.t()

Lets get rid of all non valid characters. We are assuming any language and hence using unicode syntax and not restricting ourselves to alphanumeric