Glific.Templates.InteractiveTemplates (Glific v4.3.0)

The InteractiveTemplate Context, which encapsulates and manages interactive templates

Functions

clean_template_title(interactive_content)

@spec clean_template_title(map() | nil) :: map() | nil

Cleaning interactive template title as per WhatsApp policy

copy_interactive_template(interactive_template, attrs)

@spec copy_interactive_template(Glific.Templates.InteractiveTemplate.t(), map()) ::
  {:ok, Glific.Templates.InteractiveTemplate.t()} | {:error, String.t()}

Make a copy of a interactive_template

count_interactive_templates(args)

@spec count_interactive_templates(map()) :: integer()

Return the count of interactive templates, using the same filter as list_interactives

create_interactive_template(attrs)

@spec create_interactive_template(map()) ::
  {:ok, Glific.Templates.InteractiveTemplate.t()} | {:error, Ecto.Changeset.t()}

Creates an interactive template

Examples

iex> create_interactive_template(%{field: value})
{:ok, %InteractiveTemplate{}}

iex> create_interactive_template(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_interactive_template(interactive)

@spec delete_interactive_template(Glific.Templates.InteractiveTemplate.t()) ::
  {:ok, Glific.Templates.InteractiveTemplate.t()} | {:error, Ecto.Changeset.t()}

Deletes an interactive template

Examples

iex> delete_interactive_template(interactive)
{:ok, %InteractiveTemplate{}}

iex> delete_interactive_template(interactive)
{:error, %Ecto.Changeset{}}

fetch_interactive_template(id)

@spec fetch_interactive_template(integer()) :: {:ok, any()} | {:error, any()}

Fetches a single interactive template

Returns Resource not found if the Interactive Template does not exist.

Examples

iex> fetch_interactive_template(123, 1)
  {:ok, %InteractiveTemplate{}}

iex> fetch_interactive_template(456, 1)
  {:error, ["Elixir.Glific.Templates.InteractiveTemplate", "Resource not found"]}

get_clean_interactive_content(interactive_content, arg2, arg3)

@spec get_clean_interactive_content(map(), boolean(), atom()) :: map()

Returns interactive content based on send_interactive_title field

get_interactive_body(interactive_content)

@spec get_interactive_body(map()) :: String.t()

get interactive body from the interactive content

get_interactive_template!(id)

@spec get_interactive_template!(integer()) :: Glific.Templates.InteractiveTemplate.t()

Gets a single interactive template

Raises Ecto.NoResultsError if the Interactive Template does not exist.

Examples

iex> get_interactive_template!(123)
%InteractiveTemplate{}

iex> get_interactive_template!(456)
** (Ecto.NoResultsError)

get_media(interactive_content, organization_id)

@spec get_media(map(), non_neg_integer()) :: non_neg_integer() | nil

Create a message media from interactive content and return id

get_translations(interactive_template, language_id)

@spec get_translations(Glific.Templates.InteractiveTemplate.t(), non_neg_integer()) ::
  map()

Fetch for translation of interactive message

list_interactives(args)

@spec list_interactives(map()) :: [Glific.Templates.InteractiveTemplate.t()]

Returns the list of interactive templates

Examples

iex> list_interactives()
[%InteractiveTemplate{}, ...]

translated_content(interactive_template, language_id)

@spec translated_content(Glific.Templates.InteractiveTemplate.t(), non_neg_integer()) ::
  map() | nil

Get translated interative template content

update_interactive_template(interactive, attrs)

@spec update_interactive_template(Glific.Templates.InteractiveTemplate.t(), map()) ::
  {:ok, Glific.Templates.InteractiveTemplate.t()} | {:error, Ecto.Changeset.t()}

Updates an interactive template

Examples

iex> update_interactive_template(interactive, %{field: new_value})
{:ok, %InteractiveTemplate{}}

iex> update_interactive_template(interactive, %{field: bad_value})
{:error, %Ecto.Changeset{}}