Glific.Triggers (Glific v4.6.8)

The trigger manager for all the trigger system that starts flows within Glific

Functions

count_triggers(args)

@spec count_triggers(map()) :: integer()

Return the count of triggers, using the same filter as list_triggers

create_trigger(attrs)

@spec create_trigger(map()) ::
  {:ok, Glific.Triggers.Trigger.t()} | {:error, Ecto.Changeset.t()}

Creates a trigger.

Examples

iex> create_trigger(%{field: value})
{:ok, %Trigger{}}

iex> create_trigger(%{field: bad_value})
{:error, %Ecto.Changeset{}}

execute_triggers(organization_id, now \\ DateTime.utc_now())

@spec execute_triggers(non_neg_integer(), DateTime.t()) :: [
  Glific.Triggers.Trigger.t()
]

Periodic call to execute the triggers outstanding for the day

get_trigger!(id)

@spec get_trigger!(integer()) :: Glific.Triggers.Trigger.t()

Gets a single trigger.

Raises Ecto.NoResultsError if the Trigger does not exist.

Examples

iex> get_trigger!(123)
%Trigger{}

iex> get_trigger!(456)
** (Ecto.NoResultsError)

list_triggers(args)

@spec list_triggers(map()) :: [Glific.Triggers.Trigger.t()]

Returns the list of triggers filtered by args

Examples

iex> list_triggers()
[%Trigger{}, ...]

update_trigger(trigger, attrs)

@spec update_trigger(Glific.Triggers.Trigger.t(), map()) ::
  {:ok, Glific.Triggers.Trigger.t()} | {:error, Ecto.Changeset.t()}

Updates a trigger.

Examples

iex> update_trigger(trigger, %{field: new_value})
{:ok, %Trigger{}}

iex> update_trigger(trigger, %{field: bad_value})
{:error, %Ecto.Changeset{}}