Glific.Users (Glific v4.3.0)

The Users context.

Functions

count_users(args)

@spec count_users(map()) :: integer()

Return the count of users, using the same filter as list_users

create(params)

Callback implementation for Pow.Context.create/1.

create_user(attrs)

@spec create_user(map()) ::
  {:ok, Glific.Users.User.t()} | {:error, Ecto.Changeset.t()}

Creates a user.

Examples

iex> create_user(%{field: value})
{:ok, %User{}}

iex> create_user(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete(user)

Callback implementation for Pow.Context.delete/1.

delete_user(user)

@spec delete_user(Glific.Users.User.t()) ::
  {:ok, Glific.Users.User.t()} | {:error, Ecto.Changeset.t()}

Deletes a user.

Examples

iex> delete_user(user)
{:ok, %User{}}

iex> delete_user(user)
{:error, %Ecto.Changeset{}}

get_by(clauses)

Callback implementation for Pow.Context.get_by/1.

get_user!(id)

@spec get_user!(integer()) :: Glific.Users.User.t()

Gets a single user.

Raises Ecto.NoResultsError if the User does not exist.

Examples

iex> get_user!(123)
%User{}

iex> get_user!(456)
** (Ecto.NoResultsError)

list_users(args)

@spec list_users(map()) :: [Glific.Users.User.t()]

Returns the list of filtered users.

Examples

iex> list_users()
[%User{}, ...]

pow_authenticate(params)

pow_create(params)

pow_delete(user)

pow_get_by(clauses)

pow_update(user, params)

promote_first_user(user)

@spec promote_first_user(Glific.Users.User.t()) :: Glific.Users.User.t()

Promote the first user of the system to admin automatically. Ignore NGO or SaaS users which are automatically created

reset_user_password(user, attrs)

@spec reset_user_password(Glific.Users.User.t(), map()) ::
  {:ok, Glific.Users.User.t()} | {:error, Ecto.Changeset.t()}

Reset user password

update(user, params)

Callback implementation for Pow.Context.update/2.

update_user(user, attrs)

@spec update_user(Glific.Users.User.t(), map()) ::
  {:ok, Glific.Users.User.t()} | {:error, Ecto.Changeset.t()}

Updates a user.

Examples

iex> update_user(user, %{field: new_value})
{:ok, %User{}}

iex> update_user(user, %{field: bad_value})
{:error, %Ecto.Changeset{}}