GlificWeb.Providers.Gupshup.Controllers.MessageController (Glific v0.9.4)

Dedicated controller to handle different types of inbound message form Gupshup

Functions

audio(conn, params)

Specs

audio(Plug.Conn.t(), map()) :: Plug.Conn.t()

Callback audio gupshup image images

file(conn, params)

Specs

file(Plug.Conn.t(), map()) :: Plug.Conn.t()

Callback file gupshup image images

image(conn, params)

Specs

image(Plug.Conn.t(), map()) :: Plug.Conn.t()

Callback for gupshup image images

sticker(conn, params)

Specs

sticker(Plug.Conn.t(), map()) :: Plug.Conn.t()

Callback for gupshup sticker image

text(conn, params)

Specs

text(Plug.Conn.t(), map()) :: Plug.Conn.t()

Parse text message payload and convert that into Glific message struct

video(conn, params)

Specs

video(Plug.Conn.t(), map()) :: Plug.Conn.t()

Callback video gupshup image images