GlificWeb.Providers.Gupshup.Controllers.MessageEventController (Glific v4.3.0)

Dedicated controller to handle all the message status requests like read, delivered etc..

Functions

delivered(conn, params)

@spec delivered(Plug.Conn.t(), map()) :: Plug.Conn.t()

Message has been delivered to whats app

enqueued(conn, params)

@spec enqueued(Plug.Conn.t(), map()) :: Plug.Conn.t()

Message status when the message has been sent to gupshup

failed(conn, params)

@spec failed(Plug.Conn.t(), map()) :: Plug.Conn.t()

Message status when gupshup could not send the message to whats app

handler(conn, params, msg)

@spec handler(Plug.Conn.t(), map(), String.t()) :: Plug.Conn.t()

Default handle for all message event callbacks

read(conn, params)

@spec read(Plug.Conn.t(), map()) :: Plug.Conn.t()

Message has been read by the beneficiary

sent(conn, params)

@spec sent(Plug.Conn.t(), map()) :: Plug.Conn.t()

Message has been sent to whats app via Gupshup