GlificWeb.Resolvers.Contacts (Glific v4.3.0)

Contact Resolver which sits between the GraphQL schema and Glific Contact Context API. This layer basically stiches together one or more calls to resolve the incoming queries.

Functions

contact_history(_, args, _)

@spec contact_history(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Release a simulator contact or nil if possible for this user

contact_location(_, map1, map2)

@spec contact_location(Absinthe.Resolution.t(), %{id: integer()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Get current location of the contact

count_contact_history(_, args, _)

@spec count_contact_history(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Release a simulator contact or nil if possible for this user

count_contacts(_, args, _)

@spec count_contacts(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, integer()}

Get the count of contacts filtered by args

import_contacts(_, map, _)

@spec import_contacts(
  Absinthe.Resolution.t(),
  %{
    group_label: String.t(),
    data: String.t(),
    id: integer(),
    type: :data | :file_path | :url
  },
  %{context: map()}
) :: {:ok, any()} | {:error, any()}

Import contacts to the database

optin_contact(_, params, _)

@spec optin_contact(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Upload a contact phone as opted in

simulator_get(_, params, map)

@spec simulator_get(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Grab a simulator contact or nil if possible for this user

simulator_release(_, params, map)

@spec simulator_release(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Release a simulator contact or nil if possible for this user