GlificWeb.Resolvers.Flows (Glific v4.3.0)

Flow Resolver which sits between the GraphQL schema and Glific Flow Context API. This layer basically stiches together one or more calls to resolve the incoming queries.

Functions

broadcast_stats(_, map, _)

@spec broadcast_stats(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Get broadcast stats for a flow

copy_flow(_, map, _)

@spec copy_flow(Absinthe.Resolution.t(), %{id: integer(), input: map()}, %{
  context: map()
}) ::
  {:ok, any()} | {:error, any()}

Make a copy a flow

count_flows(_, args, _)

@spec count_flows(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, integer()}

Get the count of flows filtered by args

flow(_, map1, map2)

@spec flow(Absinthe.Resolution.t(), %{id: integer()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Get a specific flow by id

flow_get(_, map1, map2)

@spec flow_get(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Grab a flow or nil if possible for this user

flow_release(_, params, map)

@spec flow_release(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Release a flow or nil if possible for this user

flows(_, args, _)

@spec flows(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, [Glific.Flows.Flow]}

Get the list of flows

publish_flow(_, map, _)

@spec publish_flow(Absinthe.Resolution.t(), %{uuid: String.t()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Publish a flow

reset_flow_count(_, map, _)

@spec reset_flow_count(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Reset the flow counts for a specific flow

resume_contact_flow(_, map1, map2)

@spec resume_contact_flow(
  Absinthe.Resolution.t(),
  %{flow_id: integer() | String.t(), contact_id: integer(), result: map()},
  %{context: map()}
) :: {:ok, any()} | {:error, any()}

Resume a flow for a contact

start_contact_flow(_, map1, map2)

@spec start_contact_flow(
  Absinthe.Resolution.t(),
  %{flow_id: integer() | String.t(), contact_id: integer()},
  %{context: map()}
) :: {:ok, any()} | {:error, any()}

Start a flow for a contact

start_group_flow(_, map1, map2)

@spec start_group_flow(
  Absinthe.Resolution.t(),
  %{flow_id: integer(), group_id: integer()},
  %{
    context: map()
  }
) :: {:ok, any()} | {:error, any()}

Start a flow for all contacts of a group

terminate_contact_flows(_, map, context)

@spec terminate_contact_flows(
  Absinthe.Resolution.t(),
  %{contact_id: integer()},
  map()
) :: {:ok, any()} | {:error, any()}

Terminate all flows for a contact