GlificWeb.Resolvers.Groups (Glific v4.3.0)

Group Resolver which sits between the GraphQL schema and Glific Group Context API. This layer basically stiches together one or more calls to resolve the incoming queries.

Functions

contacts_count(_, params, _)

@spec contacts_count(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, integer()}

Get count of group contacts

count_groups(_, args, _)

@spec count_groups(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, integer()}

Get the count of groups filtered by args

create_contact_group(_, map, _)

@spec create_contact_group(Absinthe.Resolution.t(), %{input: map()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Creates an contact group entry

create_group(_, map, _)

@spec create_group(Absinthe.Resolution.t(), %{input: map()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Creates an group

create_user_group(_, map, _)

@spec create_user_group(Absinthe.Resolution.t(), %{input: map()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Creates an user group entry

delete_group(_, map1, map2)

@spec delete_group(Absinthe.Resolution.t(), %{id: integer()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Deletes an group

group(_, map, context)

@spec group(Absinthe.Resolution.t(), %{id: integer()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Get a specific group by id

group_info(_, map, _)

@spec group_info(Absinthe.Resolution.t(), %{id: integer()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Get group info by id

groups(_, args, _)

@spec groups(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, [Glific.Groups.Group]}

Get the list of groups filtered by args

organization_groups(_, args, _)

@spec organization_groups(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, [Glific.Groups.Group]}

Get the list of organizations groups filtered by args

update_group(_, map1, map2)

@spec update_group(Absinthe.Resolution.t(), %{id: integer(), input: map()}, %{
  context: map()
}) ::
  {:ok, any()} | {:error, any()}

Updates an group

users_count(_, params, _)

@spec users_count(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, integer()}

Get count of group users