GlificWeb.Resolvers.Partners (Glific v4.3.0)

Partners Resolver which sits between the GraphQL schema and Glific Partners Context API. This layer basically stiches together one or more calls to resolve the incoming queries.

Functions

bspbalance(_, _, map)

@spec bspbalance(Absinthe.Resolution.t(), %{id: integer()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Get a specific bsp balance by organization id

count_organizations(_, args, _)

@spec count_organizations(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, integer()}

Get the count of organizations filtered by args

count_providers(_, args, _)

@spec count_providers(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, integer()}

Get the count of providers filtered by args

create_credential(_, map1, map2)

@spec create_credential(Absinthe.Resolution.t(), %{input: map()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Creates an organization's credential

create_organization(_, map, _)

@spec create_organization(Absinthe.Resolution.t(), %{input: map()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Creates an organization

create_provider(_, map, _)

@spec create_provider(Absinthe.Resolution.t(), %{input: map()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Creates a provider

credential(_, map1, map2)

@spec credential(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Get organization's credential by shorcode/service

delete_inactive_organization(_, map, _)

@spec delete_inactive_organization(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Delete an inactive organization

delete_organization(_, map, _)

@spec delete_organization(Absinthe.Resolution.t(), %{id: integer()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Deletes an organization

delete_provider(_, map, _)

@spec delete_provider(Absinthe.Resolution.t(), %{id: integer()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Deletes a provider

organization(_, arg2, arg3)

@spec organization(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Get a specific organization by id

organization_services(_, _, map)

@spec organization_services(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, map()}

Get the organizations services

organizations(_, args, _)

@spec organizations(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Get the list of organizations filtered by args

provider(_, map, _)

@spec provider(Absinthe.Resolution.t(), %{id: integer()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Get a specific provider by id

providers(_, args, _)

@spec providers(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Get the list of providers

reset_organization(_, map, _)

@spec reset_organization(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok | :error, String.t()}

Resets table and some columns of an organization

update_credential(_, map1, map2)

@spec update_credential(Absinthe.Resolution.t(), %{id: integer(), input: map()}, %{
  context: map()
}) ::
  {:ok, any()} | {:error, any()}

Updates an organization's credential

update_organization(_, map, _)

@spec update_organization(Absinthe.Resolution.t(), %{id: integer(), input: map()}, %{
  context: map()
}) ::
  {:ok, any()} | {:error, any()}

Updates an organization

update_organization_status(_, map, _)

@spec update_organization_status(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Updates an organization status is_active/is_approved. We will add checks to validate approval and activation

update_provider(_, map, _)

@spec update_provider(Absinthe.Resolution.t(), %{id: integer(), input: map()}, %{
  context: map()
}) ::
  {:ok, any()} | {:error, any()}

Updates a provider