Glific v0.2.0 GlificWeb.Resolvers.Settings

Settings Resolver which sits between the GraphQL schema and Glific Settings Context API. This layer basically stiches together one or more calls to resolve the incoming queries.

Functions

count_languages(_, args, _)

Specs

count_languages(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, integer()}

Get the count of languages filtered by args

create_language(_, map, _)

Specs

create_language(Absinthe.Resolution.t(), %{input: map()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Create a new language. Since language is a basic system data type, this operation is an upsert

language(_, map, _)

Specs

language(Absinthe.Resolution.t(), %{id: integer()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Get a specific language by id

languages(_, args, _)

Specs

languages(Absinthe.Resolution.t(), %{required(atom()) => any()}, %{
  context: map()
}) :: {:ok, any()} | {:error, any()}

Get the list of languages filtered by args

search(_, map, _)

Specs

search(Absinthe.Resolution.t(), %{matching: String.t()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Get the list of objects in the database that match the term

update_language(_, map, _)

Specs

update_language(Absinthe.Resolution.t(), %{id: integer(), input: map()}, %{
  context: map()
}) :: {:ok, any()} | {:error, any()}

Update language data fields