GlificWeb.Resolvers.Tags (Glific v0.9.4)

Tag Resolver which sits between the GraphQL schema and Glific Tag Context API. This layer basically stiches together one or more calls to resolve the incoming queries.

Functions

count_tags(_, args, _)

Specs

count_tags(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, integer()}

Get the count of tags filtered by args

create_template_tag(_, map, _)

Specs

create_template_tag(Absinthe.Resolution.t(), %{input: map()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Create entry for tag mapped to template

tag(_, map1, map2)

Specs

tag(Absinthe.Resolution.t(), %{id: integer()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Get a specific tag by id

tags(_, args, _)

Specs

tags(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, [Glific.Tags.Tag]}

Get the list of tags filtered by args

update_contact_tags(_, map, _)

Specs

update_contact_tags(Absinthe.Resolution.t(), %{input: map()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Creates and/or deletes a list of contact tags, each tag attached to the same contact

update_template_tags(_, map, _)

Specs

update_template_tags(Absinthe.Resolution.t(), %{input: map()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Creates and/or deletes a list of template tags, each tag attached to the same template