GlificWeb.Resolvers.Templates (Glific v4.3.0)

Templates Resolver which sits between the GraphQL schema and Glific Templates Context API. This layer basically stiches together one or more calls to resolve the incoming queries.

Functions

count_session_templates(_, args, _)

@spec count_session_templates(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, integer()}

Get the count of session templates filtered by args

create_template_from_message(_, params, _)

@spec create_template_from_message(
  Absinthe.Resolution.t(),
  %{message_id: integer(), input: map()},
  %{context: map()}
) :: {:ok, any()} | {:error, any()}

Converting a message to message template

import_templates(_, map1, map2)

@spec import_templates(Absinthe.Resolution.t(), %{data: String.t()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Import pre approved templates

session_template(_, map1, map2)

@spec session_template(Absinthe.Resolution.t(), %{id: integer()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Get a specific session template by id

session_templates(_, args, _)

@spec session_templates(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Get the list of session templates filtered by args