GlificWeb.Router.Helpers (Glific v4.3.0)

Module with named helpers generated from GlificWeb.Router.

Functions

api_v1_onboard_path(conn_or_endpoint, action)

api_v1_onboard_path(conn_or_endpoint, action, params)

api_v1_onboard_url(conn_or_endpoint, action)

api_v1_onboard_url(conn_or_endpoint, action, params)

api_v1_registration_path(conn_or_endpoint, action)

api_v1_registration_path(conn_or_endpoint, action, params)

api_v1_registration_url(conn_or_endpoint, action)

api_v1_registration_url(conn_or_endpoint, action, params)

api_v1_session_path(conn_or_endpoint, action)

api_v1_session_path(conn_or_endpoint, action, params)

api_v1_session_url(conn_or_endpoint, action)

api_v1_session_url(conn_or_endpoint, action, params)

exotel_path(conn_or_endpoint, action)

exotel_path(conn_or_endpoint, action, params)

exotel_url(conn_or_endpoint, action)

exotel_url(conn_or_endpoint, action, params)

flow_editor_path(conn_or_endpoint, action)

flow_editor_path(conn_or_endpoint, action, params)

flow_editor_path(conn_or_endpoint, action, vars, params)

flow_editor_url(conn_or_endpoint, action)

flow_editor_url(conn_or_endpoint, action, params)

flow_editor_url(conn_or_endpoint, action, vars, params)

live_dashboard_path(conn_or_endpoint, action)

live_dashboard_path(conn_or_endpoint, action, params)

live_dashboard_path(conn_or_endpoint, action, page, params)

live_dashboard_path(conn_or_endpoint, action, node, page, params)

live_dashboard_url(conn_or_endpoint, action)

live_dashboard_url(conn_or_endpoint, action, params)

live_dashboard_url(conn_or_endpoint, action, page, params)

live_dashboard_url(conn_or_endpoint, action, node, page, params)

live_path(conn_or_endpoint, action)

live_path(conn_or_endpoint, action, params)

live_url(conn_or_endpoint, action)

live_url(conn_or_endpoint, action, params)

oban_dashboard_path(conn_or_endpoint, action)

oban_dashboard_path(conn_or_endpoint, action, params)

oban_dashboard_path(conn_or_endpoint, action, page, params)

oban_dashboard_path(conn_or_endpoint, action, page, id, params)

oban_dashboard_url(conn_or_endpoint, action)

oban_dashboard_url(conn_or_endpoint, action, params)

oban_dashboard_url(conn_or_endpoint, action, page, params)

oban_dashboard_url(conn_or_endpoint, action, page, id, params)

path(data, path)

Generates the path information including any necessary prefix.

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

stripe_path(conn_or_endpoint, action)

stripe_path(conn_or_endpoint, action, params)

stripe_url(conn_or_endpoint, action)

stripe_url(conn_or_endpoint, action, params)

url(data)

Generates the connection/endpoint base URL without any path information.

webhook_path(conn_or_endpoint, action)

webhook_path(conn_or_endpoint, action, params)

webhook_url(conn_or_endpoint, action)

webhook_url(conn_or_endpoint, action, params)