Glific v0.2.0 GlificWeb.Schema

This is the container for the top level Absinthe GraphQL schema which encapsulates the entire Glific Public API. This file is primarily a container and pulls in the relevant information for data type specific files.

Functions

context(context)

Specs

context(map()) :: map()

Used to set some values in the context that we may need in order to run. For now we are just using it for Dataloader perspectives.

I think we will be storing authentication and current user in the context map in future releases. We have already started storing current user info in the context map.

middleware(middleware, field, arg3)

Specs

middleware(
  [Absinthe.Middleware.spec(), ...],
  Absinthe.Type.Field.t(),
  Absinthe.Type.Object.t()
) :: [Absinthe.Middleware.spec(), ...]

Used to apply middleware on all or a group of fields based on pattern matching.

It is passed the existing middleware for a field, the field itself, and the object that the field is a part of.

plugins()

Specs

plugins() :: [Absinthe.Plugin.t()]

Used to define the list of plugins to run before and after resolution.