GlificWeb.Schema.Middleware.QueryErrors (Glific v4.3.0)

Implementing middleware functions to transform errors from Elixir and friends into a format consumable and displayable to the API user. This version is specifically for queries.

Functions

call(res, _)

@spec call(Absinthe.Resolution.t(), term()) :: Absinthe.Resolution.t()

This is the main middleware callback.

It receives an %Absinthe.Resolution{} struct and it needs to return an %Absinthe.Resolution{} struct. The second argument will be whatever value was passed to the middleware call that setup the middleware.