Glific (Glific v0.9.4)

Glific keeps the contexts that define your domain and business logic.

Contexts are also responsible for managing your data, regardless if it comes from the database, an external API or others.

For now we'll keep some commonly used functions here, until we need a new file

Functions

atomize_keys(map)

Specs

atomize_keys(map()) :: map()

Convert map string keys to :atom keys

go_back_time(go_back, time \\ DateTime.utc_now(), unit \\ :hour)

Specs

go_back_time(integer(), DateTime.t(), atom()) :: DateTime.t()

Return a time object where you go back x units. We introduce the notion of hour and minute

in_past_time(time, units \\ :hours, back \\ 24)

Specs

in_past_time(DateTime.t(), atom(), integer()) :: boolean()

See if the current time is within the past time units

parse_maybe_integer(value)

Specs

parse_maybe_integer(String.t() | integer()) :: {:ok, integer()} | :error

Wrapper to return :ok/:error when parsing strings to potential integers

signature(organization_id, body, timestamp)

Specs

signature(non_neg_integer(), String.t(), non_neg_integer()) :: String.t()

Compute the signature at a specific time for the body

stacktrace()

Specs

stacktrace() :: :ok

easy way for glific developers to get a stacktrace when debugging

string_clean(str)

Specs

string_clean(String.t() | nil) :: String.t() | nil

Lets get rid of all non valid characters. We are assuming any language and hence using unicode syntax and not restricting ourselves to alphanumeric

suspicious_code(code)

Specs

suspicious_code(String.t()) :: boolean()

Really simple function to ensure folks do not add Repo and/or IO calls to an EEx snippet. This is an extremely short term fix to avoid shooting ourselves in the foot, but we should move to lua for flows scripting in the near future

Note that folks can potentially find other ways to access the same modules, so this by no means should be considered remotely secure

validate_shortcode(changeset)

Specs

validate_shortcode(Ecto.Changeset.t()) ::
  Ecto.Changeset.t() | Ecto.Changeset.t()

Validates inputed shortcode, if shortcode is invalid it returns message that the shortcode is invalid along with the valid shortcode.