Glific v0.2.0 Glific

Glific keeps the contexts that define your domain and business logic.

Contexts are also responsible for managing your data, regardless if it comes from the database, an external API or others.

For now we'll keep some commonly used functions here, until we need a new file

Functions

parse_maybe_integer(value)

Specs

parse_maybe_integer(String.t() | integer()) :: {:ok, integer()} | :error

Wrapper to return :ok/:error when parsing strings to potential integers

string_clean(str)

Specs

string_clean(String.t() | nil) :: String.t() | nil

Lets get rid of all non valid characters. We are assuming any language and hence using unicode syntax and not restricting ourselves to alphanumeric