Glific v0.1.0 Glific.Communications.Message

The Message Communication Context, which encapsulates and manages tags and the related join tables.

Functions

handle_error_response(response, message)

Specs

handle_error_response(Tesla.Env.t(), Glific.Messages.Message.t()) ::
  {:error, String.t()}

Callback in case of any error while sending the message

handle_success_response(response, message)

Specs

handle_success_response(Tesla.Env.t(), Glific.Messages.Message.t()) ::
  {:ok, Glific.Messages.Message.t()}

Callback when message send succsully

receive_message(message_params, type \\ :text)

Specs

receive_message(map(), atom()) :: {:ok} | {:error, String.t()}

Callback when we receive a message from whats app

send_message(message)

Specs

send_message(Glific.Messages.Message.t()) ::
  {:ok, Glific.Messages.Message.t()} | {:error, String.t()}

Send message to receiver using define provider.

update_provider_status(provider_message_id, provider_status)

Specs

update_provider_status(String.t(), atom()) :: {:ok, Glific.Messages.Message.t()}

Callback to update the provider status for a message