Glific.Contacts (Glific v0.9.4)

The Contacts context.

Types

option()

Specs

option() ::
  {:method, Tesla.Env.method()}
  | {:url, Tesla.Env.url()}
  | {:query, Tesla.Env.query()}
  | {:headers, Tesla.Env.headers()}
  | {:body, Tesla.Env.body()}
  | {:opts, Tesla.Env.opts()}

Functions

add_permission(query, user)

Specs

add_permission(Ecto.Query.t(), Glific.Users.User.t()) :: Ecto.Query.t()

Add permissioning specific to groups, in this case we want to restrict the visibility of groups that the user can see

can_send_message_to?(contact)

Specs

can_send_message_to?(Glific.Contacts.Contact.t()) :: boolean()

Check if we can send a message to the contact

can_send_message_to?(contact, is_hsm)

Specs

can_send_message_to?(Glific.Contacts.Contact.t(), boolean()) :: boolean()

Check if we can send a session message to the contact

change_contact(contact, attrs \\ %{})

Specs

change_contact(Glific.Contacts.Contact.t(), map()) :: Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking contact changes.

Examples

iex> change_contact(contact)
%Ecto.Changeset{data: %Contact{}}

contact_location(contact)

Specs

contact_location(Glific.Contacts.Contact.t()) ::
  {:ok, Glific.Contacts.Location.t()}

Get contact's current location

contact_opted_in(phone, organization_id, utc_time)

Specs

contact_opted_in(String.t(), non_neg_integer(), DateTime.t()) ::
  {:ok, Glific.Contacts.Contact.t()} | {:error, Ecto.Changeset.t()}

Update DB fields when contact opted in and ignore if it's blocked

contact_opted_out(phone, organization_id, utc_time)

Specs

contact_opted_out(String.t(), non_neg_integer(), DateTime.t()) :: {:ok}

Update DB fields when contact opted out

count_contacts(args)

Specs

count_contacts(map()) :: integer()

Return the count of contacts, using the same filter as list_contacts

create_contact(attrs)

Specs

create_contact(map()) ::
  {:ok, Glific.Contacts.Contact.t()} | {:error, Ecto.Changeset.t()}

Creates a contact.

Examples

iex> create_contact(%{field: value})
{:ok, %Contact{}}

iex> create_contact(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_location(attrs \\ %{})

Specs

create_location(map()) ::
  {:ok, Glific.Contacts.Location.t()} | {:error, Ecto.Changeset.t()}

Creates a location.

Examples

iex> Glific.Contacts.create_location(%{name: value})
{:ok, %Glific.Contacts.Location{}}

iex> Glific.Contacts.create_location(%{bad_field: bad_value})
{:error, %Ecto.Changeset{}}

delete(client, url, opts)

Specs

delete(Tesla.Env.client(), Tesla.Env.url(), [option()]) :: Tesla.Env.result()

Perform a DELETE request.

See request/1 or request/2 for options definition.

delete("/users")
delete("/users", query: [scope: "admin"])
delete(client, "/users")
delete(client, "/users", query: [scope: "admin"])
delete(client, "/users", body: %{name: "Jon"})

delete!(client, url, opts)

Specs

delete!(Tesla.Env.client(), Tesla.Env.url(), [option()]) ::
  Tesla.Env.t() | no_return()

Perform a DELETE request.

See request!/1 or request!/2 for options definition.

delete!("/users")
delete!("/users", query: [scope: "admin"])
delete!(client, "/users")
delete!(client, "/users", query: [scope: "admin"])
delete!(client, "/users", body: %{name: "Jon"})

delete_contact(contact)

Specs

delete_contact(Glific.Contacts.Contact.t()) ::
  {:ok, Glific.Contacts.Contact.t()} | {:error, Ecto.Changeset.t()}

Deletes a contact.

Examples

iex> delete_contact(contact)
{:ok, %Contact{}}

iex> delete_contact(contact)
{:error, %Ecto.Changeset{}}

get(client, url, opts)

Specs

get(Tesla.Env.client(), Tesla.Env.url(), [option()]) :: Tesla.Env.result()

Perform a GET request.

See request/1 or request/2 for options definition.

get("/users")
get("/users", query: [scope: "admin"])
get(client, "/users")
get(client, "/users", query: [scope: "admin"])
get(client, "/users", body: %{name: "Jon"})

get!(client, url, opts)

Specs

get!(Tesla.Env.client(), Tesla.Env.url(), [option()]) ::
  Tesla.Env.t() | no_return()

Perform a GET request.

See request!/1 or request!/2 for options definition.

get!("/users")
get!("/users", query: [scope: "admin"])
get!(client, "/users")
get!(client, "/users", query: [scope: "admin"])
get!(client, "/users", body: %{name: "Jon"})

get_contact!(id)

Specs

get_contact!(integer()) :: Glific.Contacts.Contact.t()

Gets a single contact.

Raises Ecto.NoResultsError if the Contact does not exist.

Examples

iex> get_contact!(123)
%Contact{}

iex> get_contact!(456)
** (Ecto.NoResultsError)

get_contact_field_map(contact_id)

Specs

get_contact_field_map(integer()) :: map()

Convert contact field to map for variable substitution

head(client, url, opts)

Specs

head(Tesla.Env.client(), Tesla.Env.url(), [option()]) :: Tesla.Env.result()

Perform a HEAD request.

See request/1 or request/2 for options definition.

head("/users")
head("/users", query: [scope: "admin"])
head(client, "/users")
head(client, "/users", query: [scope: "admin"])
head(client, "/users", body: %{name: "Jon"})

head!(client, url, opts)

Specs

head!(Tesla.Env.client(), Tesla.Env.url(), [option()]) ::
  Tesla.Env.t() | no_return()

Perform a HEAD request.

See request!/1 or request!/2 for options definition.

head!("/users")
head!("/users", query: [scope: "admin"])
head!(client, "/users")
head!(client, "/users", query: [scope: "admin"])
head!(client, "/users", body: %{name: "Jon"})

is_contact_blocked?(phone, organization_id)

Specs

is_contact_blocked?(String.t(), non_neg_integer()) :: boolean()

check if contact is blocked or not

is_new_contact(contact_id)

Specs

is_new_contact(integer()) :: boolean()

Check if this contact id is a new contact. In general, we should always retrive as little as possible from the DB

list_contacts(args)

Specs

list_contacts(map()) :: [Glific.Contacts.Contact.t()]

Returns the list of contacts.

Examples

iex> list_contacts()
[%Contact{}, ...]

Get the list of contacts filtered by various search options Include contacts only if within list of groups Include contacts only if have list of tags

list_user_contacts(args \\ %{})

Specs

list_user_contacts(map()) :: [Glific.Contacts.Contact.t()]

Return the list of contacts who are also users

maybe_create_contact(sender)

Specs

maybe_create_contact(map()) ::
  {:ok, Glific.Contacts.Contact.t()} | {:error, Ecto.Changeset.t()}

This function is called by the messaging framework for all incoming messages, hence might be a good candidate to maintain a contact level cache at some point

We use a fetch followed by create, to avoid the explosion in the id namespace

optin_contact(attrs)

Specs

optin_contact(map()) ::
  {:ok, Glific.Contacts.Contact.t()}
  | {:error, Ecto.Changeset.t()}
  | {:error, String.t()}

Upload a contact phone as opted in

options(client, url, opts)

Specs

options(Tesla.Env.client(), Tesla.Env.url(), [option()]) :: Tesla.Env.result()

Perform a OPTIONS request.

See request/1 or request/2 for options definition.

options("/users")
options("/users", query: [scope: "admin"])
options(client, "/users")
options(client, "/users", query: [scope: "admin"])
options(client, "/users", body: %{name: "Jon"})

options!(client, url, opts)

Specs

options!(Tesla.Env.client(), Tesla.Env.url(), [option()]) ::
  Tesla.Env.t() | no_return()

Perform a OPTIONS request.

See request!/1 or request!/2 for options definition.

options!("/users")
options!("/users", query: [scope: "admin"])
options!(client, "/users")
options!(client, "/users", query: [scope: "admin"])
options!(client, "/users", body: %{name: "Jon"})

patch(client, url, body, opts)

Specs

patch(Tesla.Env.client(), Tesla.Env.url(), Tesla.Env.body(), [option()]) ::
  Tesla.Env.result()

Perform a PATCH request.

See request/1 or request/2 for options definition.

patch("/users", %{name: "Jon"})
patch("/users", %{name: "Jon"}, query: [scope: "admin"])
patch(client, "/users", %{name: "Jon"})
patch(client, "/users", %{name: "Jon"}, query: [scope: "admin"])

patch!(client, url, body, opts)

Specs

patch!(Tesla.Env.client(), Tesla.Env.url(), Tesla.Env.body(), [option()]) ::
  Tesla.Env.t() | no_return()

Perform a PATCH request.

See request!/1 or request!/2 for options definition.

patch!("/users", %{name: "Jon"})
patch!("/users", %{name: "Jon"}, query: [scope: "admin"])
patch!(client, "/users", %{name: "Jon"})
patch!(client, "/users", %{name: "Jon"}, query: [scope: "admin"])

post(client, url, body, opts)

Specs

post(Tesla.Env.client(), Tesla.Env.url(), Tesla.Env.body(), [option()]) ::
  Tesla.Env.result()

Perform a POST request.

See request/1 or request/2 for options definition.

post("/users", %{name: "Jon"})
post("/users", %{name: "Jon"}, query: [scope: "admin"])
post(client, "/users", %{name: "Jon"})
post(client, "/users", %{name: "Jon"}, query: [scope: "admin"])

post!(client, url, body, opts)

Specs

post!(Tesla.Env.client(), Tesla.Env.url(), Tesla.Env.body(), [option()]) ::
  Tesla.Env.t() | no_return()

Perform a POST request.

See request!/1 or request!/2 for options definition.

post!("/users", %{name: "Jon"})
post!("/users", %{name: "Jon"}, query: [scope: "admin"])
post!(client, "/users", %{name: "Jon"})
post!(client, "/users", %{name: "Jon"}, query: [scope: "admin"])

put(client, url, body, opts)

Specs

put(Tesla.Env.client(), Tesla.Env.url(), Tesla.Env.body(), [option()]) ::
  Tesla.Env.result()

Perform a PUT request.

See request/1 or request/2 for options definition.

put("/users", %{name: "Jon"})
put("/users", %{name: "Jon"}, query: [scope: "admin"])
put(client, "/users", %{name: "Jon"})
put(client, "/users", %{name: "Jon"}, query: [scope: "admin"])

put!(client, url, body, opts)

Specs

put!(Tesla.Env.client(), Tesla.Env.url(), Tesla.Env.body(), [option()]) ::
  Tesla.Env.t() | no_return()

Perform a PUT request.

See request!/1 or request!/2 for options definition.

put!("/users", %{name: "Jon"})
put!("/users", %{name: "Jon"}, query: [scope: "admin"])
put!(client, "/users", %{name: "Jon"})
put!(client, "/users", %{name: "Jon"}, query: [scope: "admin"])

request(client \\ %Tesla.Client{}, options)

Specs

request(Tesla.Env.client(), [option()]) :: Tesla.Env.result()

Perform a request.

Options

:method - the request method, one of [:head, :get, :delete, :trace, :options, :post, :put, :patch]

:url - either full url e.g. "http://example.com/some/path" or just "/some/path" if using Tesla.Middleware.BaseUrl

:query - a keyword list of query params, e.g. [page: 1, per_page: 100]

:headers - a keyworld list of headers, e.g. [{"content-type", "text/plain"}]

:body - depends on used middleware:

by default it can be a binary

if using e.g. JSON encoding middleware it can be a nested map

if adapter supports it it can be a Stream with any of the above

:opts - custom, per-request middleware or adapter options

Examples

ExampleApi.request(method: :get, url: "/users/path")

# use shortcut methods
ExampleApi.get("/users/1")
ExampleApi.post(client, "/users", %{name: "Jon"})

request!(client \\ %Tesla.Client{}, options)

Specs

request!(Tesla.Env.client(), [option()]) :: Tesla.Env.t() | no_return()

Perform request and raise in case of error.

This is similar to request/2 behaviour from Tesla 0.x

See request/2 for list of available options.

set_session_status(contact, status)

Specs

set_session_status(Glific.Contacts.Contact.t() | [non_neg_integer()], atom()) ::
  {:ok, Glific.Contacts.Contact.t()} | {:error, Ecto.Changeset.t()} | :ok

Set session status for opted in and opted out contacts

trace(client, url, opts)

Specs

trace(Tesla.Env.client(), Tesla.Env.url(), [option()]) :: Tesla.Env.result()

Perform a TRACE request.

See request/1 or request/2 for options definition.

trace("/users")
trace("/users", query: [scope: "admin"])
trace(client, "/users")
trace(client, "/users", query: [scope: "admin"])
trace(client, "/users", body: %{name: "Jon"})

trace!(client, url, opts)

Specs

trace!(Tesla.Env.client(), Tesla.Env.url(), [option()]) ::
  Tesla.Env.t() | no_return()

Perform a TRACE request.

See request!/1 or request!/2 for options definition.

trace!("/users")
trace!("/users", query: [scope: "admin"])
trace!(client, "/users")
trace!(client, "/users", query: [scope: "admin"])
trace!(client, "/users", body: %{name: "Jon"})

update_contact(contact, attrs)

Specs

update_contact(Glific.Contacts.Contact.t(), map()) ::
  {:ok, Glific.Contacts.Contact.t()} | {:error, Ecto.Changeset.t()}

Updates a contact.

Examples

iex> update_contact(contact, %{field: new_value})
{:ok, %Contact{}}

iex> update_contact(contact, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_contact_status(organization_id, args)

Specs

update_contact_status(non_neg_integer(), map()) :: :ok

Invoked from cron jobs to mass update the status of contacts belonging to a specific organization

In this case, if we can, we might want to do it across the entire DB since the update is across all organizations. The main issue might be the row level security of postgres and how it ties in. For now, lets stick to per organization

upsert(attrs)

Specs

upsert(map()) :: {:ok, Glific.Contacts.Contact.t()}

Gets or Creates a Contact based on the unique indexes in the table. If there is a match it returns the existing contact, else it creates a new one