Glific v0.2.0 Glific.Contacts

The Contacts context.

Functions

can_send_message_to?(contact)

Specs

can_send_message_to?(Glific.Contacts.Contact.t()) :: boolean()

Check if we can send a message to the contact

can_send_message_to?(contact, is_hsm)

Specs

can_send_message_to?(Glific.Contacts.Contact.t(), boolean()) :: boolean()

Check if we can send a session message to the contact

change_contact(contact, attrs \\ %{})

Specs

change_contact(Glific.Contacts.Contact.t(), map()) :: Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking contact changes.

Examples

iex> change_contact(contact)
%Ecto.Changeset{data: %Contact{}}

contact_location(contact)

Specs

contact_location(Glific.Contacts.Contact.t()) ::
  {:ok, Glific.Contacts.Location.t()}

Get contact's current location

contact_opted_in(phone, utc_time)

Specs

contact_opted_in(String.t(), DateTime.t()) :: {:ok}

Update DB fields when contact opted in

contact_opted_out(phone, utc_time)

Specs

contact_opted_out(String.t(), DateTime.t()) :: {:ok}

Update DB fields when contact opted out

count_contacts(args \\ %{})

Specs

count_contacts(map()) :: integer()

Return the count of contacts, using the same filter as list_contacts

create_contact(attrs \\ %{})

Specs

create_contact(map()) ::
  {:ok, Glific.Contacts.Contact.t()} | {:error, Ecto.Changeset.t()}

Creates a contact.

Examples

iex> create_contact(%{field: value})
{:ok, %Contact{}}

iex> create_contact(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_location(attrs \\ %{})

Specs

create_location(map()) ::
  {:ok, Glific.Contacts.Location.t()} | {:error, Ecto.Changeset.t()}

Creates a location.

Examples

iex> Glific.Contacts.create_location(%{name: value})
{:ok, %Glific.Contacts.Location{}}

iex> Glific.Contacts.create_location(%{bad_field: bad_value})
{:error, %Ecto.Changeset{}}

delete_contact(contact)

Specs

delete_contact(Glific.Contacts.Contact.t()) ::
  {:ok, Glific.Contacts.Contact.t()} | {:error, Ecto.Changeset.t()}

Deletes a contact.

Examples

iex> delete_contact(contact)
{:ok, %Contact{}}

iex> delete_contact(contact)
{:error, %Ecto.Changeset{}}

get_contact!(id)

Specs

get_contact!(integer()) :: Glific.Contacts.Contact.t()

Gets a single contact.

Raises Ecto.NoResultsError if the Contact does not exist.

Examples

iex> get_contact!(123)
%Contact{}

iex> get_contact!(456)
** (Ecto.NoResultsError)

is_new_contact(contact_id)

Specs

is_new_contact(integer()) :: boolean()

Check if this contact id is a new conatct

list_contacts(args \\ %{})

Specs

list_contacts(map()) :: [Glific.Contacts.Contact.t()]

Returns the list of contacts.

Examples

iex> list_contacts()
[%Contact{}, ...]

update_contact(contact, attrs)

Specs

update_contact(Glific.Contacts.Contact.t(), map()) ::
  {:ok, Glific.Contacts.Contact.t()} | {:error, Ecto.Changeset.t()}

Updates a contact.

Examples

iex> update_contact(contact, %{field: new_value})
{:ok, %Contact{}}

iex> update_contact(contact, %{field: bad_value})
{:error, %Ecto.Changeset{}}

upsert(attrs)

Specs

upsert(map()) :: {:ok, Glific.Contacts.Contact.t()}

Gets or Creates a Contact based on the unique indexes in the table. If there is a match it returns the existing contact, else it creates a new one