Glific v0.2.0 Glific.Enums

The Enum provides a location for all enum related macros. All the constants that Ecto/Elixir used are exposed here as macros, so other files can invoke them as simple functions

Functions

api_status_const()
(macro)

Lets ensure that the constants evaluate to some expected values

Examples

iex> Glific.Enums.api_status_const() [:ok, :error]

iex> Glific.Enums.contact_status_const() [:failed, :invalid, :processing, :valid]

iex> Glific.Enums.flow_case_const() [:has_any_word]

iex> Glific.Enums.flow_router_const() [:switch]

iex> Glific.Enums.flow_action_type_const() [:enter_flow, :send_msg, :set_contact_language, :wait_for_response, :set_contact_field]

iex> Glific.Enums.flow_type_const() [:message]

iex> Glific.Enums.message_flow_const() [:inbound, :outbound]

iex> Glific.Enums.message_status_const() [:sent, :delivered, :enqueued, :error, :read, :received, :contact_opt_out]

iex> Glific.Enums.message_type_const() [:audio, :contact, :document, :hsm, :image, :location, :text, :video]

iex> Glific.Enums.question_type_const() [:text, :numeric, :date]

iex> Glific.Enums.sort_order_const() [:asc, :desc]

We also test the ecto enums in this file, since they exist outside a module

iex> Glific.Enums.APIStatus.enum_map() Glific.Enums.api_status_const()

iex> Glific.Enums.ContactStatus.enum_map() Glific.Enums.contact_status_const()

iex> Glific.Enums.FlowCase.enum_map() Glific.Enums.flow_case_const()

iex> Glific.Enums.FlowRouter.enum_map() Glific.Enums.flow_router_const()

iex> Glific.Enums.FlowActionType.enum_map() Glific.Enums.flow_action_type_const()

iex> Glific.Enums.FlowType.enum_map() Glific.Enums.flow_type_const()

iex> Glific.Enums.MessageFlow.enum_map() Glific.Enums.message_flow_const()

iex> Glific.Enums.MessageStatus.enum_map() Glific.Enums.message_status_const()

iex> Glific.Enums.MessageType.enum_map() Glific.Enums.message_type_const()

iex> Glific.Enums.QuestionType.enum_map() Glific.Enums.question_type_const()

iex> Glific.Enums.SortOrder.enum_map() Glific.Enums.sort_order_const()

contact_status_const()
(macro)

flow_action_type_const()
(macro)

flow_case_const()
(macro)

flow_router_const()
(macro)

flow_type_const()
(macro)

message_flow_const()
(macro)

message_status_const()
(macro)

message_type_const()
(macro)

question_type_const()
(macro)

sort_order_const()
(macro)