Glific.Flows (Glific v0.9.4)

The Flows context.

Functions

build_flow_objects(json, uuid_map, process_fn, object \\ nil)

Specs

build_flow_objects(map(), map(), (map(), map(), any() -> {any(), map()}), any()) ::
  {any(), map()}

A generic json traversal and building the structure for a specific flow schema which is an array of objects in the json file. Used for Node/Actions, Node/Exits, Router/Cases, and Router/Categories

change_flow(flow, attrs \\ %{})

Specs

change_flow(Glific.Flows.Flow.t(), map()) :: Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking flow changes.

Examples

iex> change_flow(flow)
%Ecto.Changeset{data: %Flow{}}

check_required_fields(json, required)

Specs

check_required_fields(map(), [atom()]) :: boolean()

Check the required fields for all flow objects. If missing, raise an exception

copy_flow(flow, attrs)

Specs

copy_flow(Glific.Flows.Flow.t(), map()) ::
  {:ok, Glific.Flows.Flow.t()} | {:error, String.t()}

Make a copy of a flow

count_flows(args)

Specs

count_flows(map()) :: integer()

Return the count of tags, using the same filter as list_tags

create_flow(attrs)

Specs

create_flow(map()) ::
  {:ok, Glific.Flows.Flow.t()} | {:error, Ecto.Changeset.t()}

Creates a flow.

Examples

iex> create_flow(%{field: value})
{:ok, %Flow{}}

iex> create_flow(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_flow_revision(definition)

Specs

create_flow_revision(map()) :: Glific.Flows.FlowRevision.t()

Save new revision for the flow

delete_flow(flow)

Specs

delete_flow(Glific.Flows.Flow.t()) ::
  {:ok, Glific.Flows.Flow.t()} | {:error, Ecto.Changeset.t()}

Deletes a flow.

Examples

iex> delete_flow(flow)
{:ok, %Flow{}}

iex> delete_flow(flow)
{:error, %Ecto.Changeset{}}

flow_activated(flow_id, contact_id, since)

Specs

flow_activated(non_neg_integer(), non_neg_integer(), DateTime.t()) :: boolean()

Check if a flow has been activated since the time sent as a parameter e.g. outofoffice will check if that flow was activated in the last 24 hours daily/weekly will check since start of day/week, etc

flow_keywords_map(organization_id)

Specs

flow_keywords_map(non_neg_integer()) :: map()

Create a map of keywords that map to flow ids for each active organization. Also cache this value including the outoffice shortcode

get_cached_flow(organization_id, key)

Specs

get_cached_flow(non_neg_integer(), {atom(), any(), String.t()}) ::
  {atom(), any()} | atom()

A helper function to interact with the Caching API and get the cached flow. It will also set the loaded flow to cache in case it does not exists.

get_flow!(id)

Specs

get_flow!(integer()) :: Glific.Flows.Flow.t()

Gets a single flow.

Raises Ecto.NoResultsError if the Flow does not exist.

Examples

iex> get_flow!(123)
%Flow{}

iex> get_flow!(456)
** (Ecto.NoResultsError)

get_flow_revision(flow_uuid, revision_id)

Specs

get_flow_revision(String.t(), String.t()) :: map()

Get specific flow revision by number

get_flow_revision_list(flow_uuid)

Specs

get_flow_revision_list(String.t()) :: %{results: list()}

Get a list of all the revisions based on a flow UUID

list_flows(args)

Specs

list_flows(map()) :: [Glific.Flows.Flow.t()]

Returns the list of flows.

Examples

iex> list_flows()
[%Flow{}, ...]

publish_flow(flow)

Specs

publish_flow(Glific.Flows.Flow.t()) :: {:ok, Glific.Flows.Flow.t()}

Update latest flow revision status as published and increment the version Update cached flow definition

start_contact_flow(flow, contact)

Specs

start_contact_flow(Glific.Flows.Flow.t(), Glific.Contacts.Contact.t()) ::
  {:ok, Glific.Flows.Flow.t()} | {:error, String.t()}

Start flow for a contact

start_group_flow(flow, group)

Specs

start_group_flow(Glific.Flows.Flow.t(), Glific.Groups.Group.t()) ::
  {:ok, Glific.Flows.Flow.t()}

Start flow for contacts of a group

update_cached_flow(flow, status)

Specs

update_cached_flow(Glific.Flows.Flow.t(), String.t()) :: {atom(), any()}

Update the cached flow from db. This typically happens when the flow definition is updated via the UI

update_flow(flow, attrs)

Specs

update_flow(Glific.Flows.Flow.t(), map()) ::
  {:ok, Glific.Flows.Flow.t()} | {:error, Ecto.Changeset.t()}

Updates a flow.

Examples

iex> update_flow(flow, %{field: new_value})
{:ok, %Flow{}}

iex> update_flow(flow, %{field: bad_value})
{:error, %Ecto.Changeset{}}