Glific v0.2.0 Glific.Flows

The Flows context.

Functions

build_flow_objects(json, uuid_map, process_fn, object \\ nil)

Specs

build_flow_objects(map(), map(), (map(), map(), any() -> {any(), map()}), any()) ::
  {any(), map()}

A generic json traversal and building the structure for a specific flow schema which is an array of objects in the json file. Used for Node/Actions, Node/Exits, Router/Cases, and Router/Categories

change_flow(flow, attrs \\ %{})

Specs

change_flow(Glific.Flows.Flow.t(), map()) :: Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking flow changes.

Examples

iex> change_flow(flow)
%Ecto.Changeset{data: %Flow{}}

check_required_fields(json, required)

Specs

check_required_fields(map(), [atom()]) :: boolean()

Check the required fields for all flow objects. If missing, raise an exception

count_flows(args \\ %{})

Specs

count_flows(map()) :: integer()

Return the count of tags, using the same filter as list_tags

create_flow(attrs \\ %{})

Specs

create_flow(map()) ::
  {:ok, Glific.Flows.Flow.t()} | {:error, Ecto.Changeset.t()}

Creates a flow.

Examples

iex> create_flow(%{field: value})
{:ok, %Flow{}}

iex> create_flow(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_flow_revision(definition)

Specs

create_flow_revision(map()) :: Glific.Flows.FlowRevision.t()

Save new revision for the flow

delete_flow(flow)

Specs

delete_flow(Glific.Flows.Flow.t()) ::
  {:ok, Glific.Flows.Flow.t()} | {:error, Ecto.Changeset.t()}

Deletes a flow.

Examples

iex> delete_flow(flow)
{:ok, %Flow{}}

iex> delete_flow(flow)
{:error, %Ecto.Changeset{}}

get_cached_flow(key, args)

Specs

get_cached_flow(any(), any()) :: {atom(), any()}

A helper function to intract with the Caching API and get the cached flow. It will also set the loaded flow to cache in case it does not exists.

get_flow!(id)

Specs

get_flow!(integer()) :: Glific.Flows.Flow.t()

Gets a single flow.

Raises Ecto.NoResultsError if the Flow does not exist.

Examples

iex> get_flow!(123)
%Flow{}

iex> get_flow!(456)
** (Ecto.NoResultsError)

get_flow_revision(flow_uuid, revision_id)

Specs

get_flow_revision(String.t(), String.t()) :: map()

Get specific flow revision by number

get_flow_revision_list(flow_uuid)

Specs

get_flow_revision_list(String.t()) :: %{results: list()}

Get a list of all the revisions based on a flow UUID

list_flows(args \\ %{})

Specs

list_flows(map()) :: [Glific.Flows.Flow.t()]

Returns the list of flows.

Examples

iex> list_flows()
[%Flow{}, ...]

update_cached_flow(flow_uuid)

Specs

update_cached_flow(Glific.Flows.Flow.t()) :: {atom(), any()}

Remove the flow from cache and add a new one.

update_flow(flow, attrs)

Specs

update_flow(Glific.Flows.Flow.t(), map()) ::
  {:ok, Glific.Flows.Flow.t()} | {:error, Ecto.Changeset.t()}

Updates a flow.

Examples

iex> update_flow(flow, %{field: new_value})
{:ok, %Flow{}}

iex> update_flow(flow, %{field: bad_value})
{:error, %Ecto.Changeset{}}