Glific v0.2.0 Glific.Flows.ContactField

Since many of the functions set/update fields in contact and related tables, lets centralize all the code here for now

Functions

add_contact_field(context, field, value, type)

Specs

add_contact_field(
  Glific.Flows.FlowContext.t(),
  String.t(),
  String.t(),
  String.t()
) :: Glific.Flows.FlowContext.t()

Add a field {key, value} to a contact. For now, all preferences are stored under the settings map, with a sub-map of preferences. We expect to get more clarity on this soon

reset_contact_fields(context)

Specs

reset_contact_fields(Glific.Flows.FlowContext.t()) ::
  Glific.Flows.FlowContext.t()

Reset the fields for a contact.