Glific.Flows.ContactSetting (Glific v4.3.0)

Since many of the functions set/update fields in contact and related tables, lets centralize all the code here for now

Functions

add_contact_preference(context, preference, value \\ true)

@spec add_contact_preference(Glific.Flows.FlowContext.t(), String.t(), boolean()) ::
  Glific.Flows.FlowContext.t()

Add a preference to a contact. For now, all preferences are stored under the settings map, with a sub-map of preferences. We expect to get more clarity on this soon

delete_contact_preference(context, preference)

@spec delete_contact_preference(Glific.Flows.FlowContext.t(), String.t()) ::
  Glific.Flows.FlowContext.t()

Delete a preference from a contact. We actually dont really delete it, we just set the value to false, and hence turn it off

get_contact_preferences(flow_context)

@spec get_contact_preferences(Glific.Flows.FlowContext.t()) :: [String.t()]

Get all the preferences for this contact

reset_contact_preference(context)

@spec reset_contact_preference(Glific.Flows.FlowContext.t()) ::
  Glific.Flows.FlowContext.t()

Reset the preferences for a contact.

set_contact_language(context, language)

@spec set_contact_language(Glific.Flows.FlowContext.t(), String.t()) ::
  Glific.Flows.FlowContext.t()

Set the language for a contact

set_contact_name(context, name)

@spec set_contact_name(Glific.Flows.FlowContext.t(), String.t()) ::
  Glific.Flows.FlowContext.t()

Set the name for a contact

set_contact_preference(context, preference)

@spec set_contact_preference(Glific.Flows.FlowContext.t(), String.t() | nil) ::
  Glific.Flows.FlowContext.t()

Wrapper function for setting the contact preference, if preference is empty, it indicates to reset the preference