Glific v0.2.0 Glific.Flows.Flow

The flow object which encapsulates the complete flow as emitted by by https://github.com/nyaruka/floweditor

Types

t()

Specs

t() :: %Glific.Flows.Flow{
  __meta__: Ecto.Schema.Metadata.t(),
  definition: map() | nil,
  flow_type: String.t() | nil,
  id: non_neg_integer() | nil,
  inserted_at: :utc_datetime | nil,
  language: Glific.Settings.Language.t() | Ecto.Association.NotLoaded.t() | nil,
  language_id: non_neg_integer() | nil,
  localization: Glific.Flows.Localization.t() | nil,
  name: String.t() | nil,
  nodes: [Glific.Flows.Node.t()] | nil,
  revisions:
    [Glific.Flows.FlowRevision.t()] | Ecto.Association.NotLoaded.t() | nil,
  shortcode: String.t() | nil,
  updated_at: :utc_datetime | nil,
  uuid: Ecto.UUID.t() | nil,
  uuid_map: map() | nil,
  version_number: String.t() | nil
}

Functions

cachex_flows(flows)

Specs

cachex_flows(map()) :: map()

Store all the flows in cachex :flows_cache. At some point, we will just use this dynamically

changeset(flow, attrs)

Specs

changeset(t(), map()) :: Ecto.Changeset.t()

Standard changeset pattern we use for all data types

context(flow, contact)

Specs

context(t(), Glific.Contacts.Contact.t()) ::
  {:ok, Glific.Flows.FlowContext.t()} | {:error, String.t()}

Build the context so we can execute the flow

get_and_cache_flows(args \\ %{})

Specs

get_and_cache_flows(map()) :: map()

Helper function for various genstage processes to set state by loading all active flows from the database and loading flows on demand

get_flow(uuid)

Specs

get_flow(Ecto.UUID.t()) :: map()

Return a flow for a specific uuid. Cache is not present in cache

get_latest_definition(flow_id)

Specs

get_latest_definition(integer()) :: map()

load the latest revision, specifically json definition from the flow_revision table. We return the clean definition back

process(json, flow)

Specs

process(map(), t()) :: t()

Process a json structure from floweditor to the Glific data types

start_sub_flow(context, uuid)

Specs

start_sub_flow(Glific.Flows.FlowContext.t(), Ecto.UUID.t()) ::
  {:ok, Glific.Flows.FlowContext.t(), [String.t()]} | {:error, String.t()}

Create a subflow of an existing flow