Glific.Flows.FlowContext (Glific v1.3.3)

When we are running a flow, we are running it in the context of a contact and/or a conversation (or other Glific data types). Let encapsulate this in a module and isolate the flow from the other aspects of Glific

Types

t()

Specs

t() :: %Glific.Flows.FlowContext{
  __meta__: Ecto.Schema.Metadata.t(),
  completed_at: :utc_datetime | nil,
  contact: Glific.Contacts.Contact.t() | Ecto.Association.NotLoaded.t() | nil,
  contact_id: non_neg_integer() | nil,
  delay: integer(),
  flow: Glific.Flows.Flow.t() | Ecto.Association.NotLoaded.t() | nil,
  flow_id: non_neg_integer() | nil,
  flow_uuid: Ecto.UUID.t() | nil,
  id: term(),
  inserted_at: :utc_datetime | nil,
  last_message: Glific.Messages.Message.t() | nil,
  node: Glific.Flows.Node.t() | nil,
  node_uuid: Ecto.UUID.t() | nil,
  organization:
    Glific.Partners.Organization.t() | Ecto.Association.NotLoaded.t() | nil,
  organization_id: non_neg_integer() | nil,
  parent: t() | Ecto.Association.NotLoaded.t() | nil,
  parent_id: non_neg_integer() | nil,
  recent_inbound: [map()] | [],
  recent_outbound: [map()] | [],
  results: map() | nil,
  status: String.t() | nil,
  updated_at: :utc_datetime | nil,
  uuid_map: map() | nil,
  wait_for_time: boolean(),
  wakeup_at: :utc_datetime | nil
}

Functions

active_context(contact_id, parent_id \\ nil)

Specs

active_context(non_neg_integer(), non_neg_integer() | nil) :: t() | nil

Check if there is an active context (i.e. with a non null, node_uuid for this contact)

changeset(context, attrs)

Specs

changeset(t(), map()) :: Ecto.Changeset.t()

Standard changeset pattern we use for all data types

create_flow_context(attrs \\ %{})

Specs

create_flow_context(map()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Create a FlowContext

delete_completed_flow_contexts(back \\ 2)

Specs

delete_completed_flow_contexts(non_neg_integer()) :: :ok

Delete all the contexts which are completed before two days

delete_old_flow_contexts(back \\ 30)

Specs

delete_old_flow_contexts(non_neg_integer()) :: :ok

Delete all the contexts which are older than 30 days

execute(context, messages)

Specs

execute(t(), [Glific.Messages.Message.t()]) ::
  {:ok | :wait, t(), [Glific.Messages.Message.t()]} | {:error, String.t()}

Execute one (or more) steps in a flow based on the message stream

get_result_value(context, value)

Specs

get_result_value(t(), String.t()) :: String.t() | nil

Retrieve the value from a results string

init_context(flow, contact, status, opts \\ [])

Specs

init_context(
  Glific.Flows.Flow.t(),
  Glific.Contacts.Contact.t(),
  String.t(),
  Keyword.t() | []
) :: {:ok, t(), [String.t()]} | {:error, String.t()}

Start a new context, if there is an existing context, blow it away

load_context(context, flow)

Specs

load_context(t(), Glific.Flows.Flow.t()) :: t()

Load the context object, given a flow object and a contact. At some point, we'll get the genserver to cache this

mark_flows_complete(contact_id, after_insert_date \\ nil)

Specs

mark_flows_complete(non_neg_integer(), DateTime.t() | nil) :: nil

Set all the flows for a specific context to be completed

match_outbound(context, body, go_back \\ 6)

Specs

match_outbound(t(), String.t(), integer()) :: integer()

Count the number of times we have sent the same message in the recent past

reset_all_contexts(context, message)

Specs

reset_all_contexts(t(), String.t()) :: t() | nil

Resets all the context for the user when we hit an error. This can potentially prevent an infinite loop from happening if flows are connected in a cycle

reset_context(context)

Specs

reset_context(t()) :: t() | nil

Resets the context and sends control back to the parent context if one exists

seed_context(flow, contact, status, opts \\ [])

Specs

seed_context(
  Glific.Flows.Flow.t(),
  Glific.Contacts.Contact.t(),
  String.t(),
  Keyword.t()
) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Seed the context and set the wakeup time as needed

set_node(context, node)

Specs

set_node(t(), Glific.Flows.Node.t()) :: t()

Set the new node for the context

step_forward(context, message)

Specs

step_forward(t(), Glific.Messages.Message.t()) ::
  {:ok, map()} | {:error, String.t()}

Given an input string, consume the input and advance the state of the context

update_recent(context, body, type)

Specs

update_recent(t(), String.t(), atom()) :: t()

Update the recent_* state as we consume or send a message

update_results(context, key, json)

Specs

update_results(t(), String.t(), map() | String.t()) :: t()

Update the contact results with each element of the json map

update_results(context, key, input, category)

Specs

update_results(t(), String.t(), String.t() | map(), String.t()) :: t()

Update the contact results state as we step through the flow

wakeup_flows(organization_id)

Specs

wakeup_flows(non_neg_integer()) :: :ok

Find all the contexts which need to be woken up and processed

wakeup_one(context, message \\ nil)

Specs

wakeup_one(t(), Glific.Messages.Message.t() | nil) ::
  {:ok, t() | nil, [String.t()]} | {:error, String.t()}

Process one context at a time that is ready to be woken