Glific v0.2.0 Glific.Flows.FlowContext

When we are running a flow, we are running it in the context of a contact and/or a conversation (or other Glific data types). Let encapsulate this in a module and isolate the flow from the other aspects of Glific

Types

t()

Specs

t() :: %Glific.Flows.FlowContext{
  __meta__: Ecto.Schema.Metadata.t(),
  completed_at: :utc_datetime | nil,
  contact: Glific.Contacts.Contact.t() | Ecto.Association.NotLoaded.t() | nil,
  contact_id: non_neg_integer() | nil,
  flow: Glific.Flows.Flow.t() | Ecto.Association.NotLoaded.t() | nil,
  flow_id: non_neg_integer() | nil,
  flow_uuid: Ecto.UUID.t() | nil,
  id: term(),
  inserted_at: :utc_datetime | nil,
  node: Glific.Flows.Node.t() | nil,
  node_uuid: Ecto.UUID.t() | nil,
  parent: t() | Ecto.Association.NotLoaded.t() | nil,
  parent_id: non_neg_integer() | nil,
  results: map() | nil,
  updated_at: :utc_datetime | nil,
  uuid_map: map() | nil,
  wakeup_at: :utc_datetime | nil
}

Functions

active_context(contact_id)

Specs

active_context(non_neg_integer()) :: t() | nil

Check if there is an active context (i.e. with a non null, node_uuid for this contact)

changeset(context, attrs)

Specs

changeset(t(), map()) :: Ecto.Changeset.t()

Standard changeset pattern we use for all data types

create_flow_context(attrs \\ %{})

Specs

create_flow_context(map()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Create a FlowContext

execute(context, messages)

Specs

execute(t(), [String.t()]) :: {:ok, t(), [String.t()]} | {:error, String.t()}

Execute one (or more) steps in a flow based on the message stream

get_result_value(context, value)

Specs

get_result_value(t(), String.t()) :: String.t() | nil

Retrieve the value from a results string

init_context(flow, contact, parent_id \\ nil)

Specs

init_context(
  Glific.Flows.Flow.t(),
  Glific.Contacts.Contact.t(),
  non_neg_integer() | nil
) :: {:ok, t(), [String.t()]} | {:error, String.t()}

Start a new context, if there is an existing context, blow it away

load_context(context, flow)

Specs

load_context(t(), Glific.Flows.Flow.t()) :: t()

Load the context object, given a flow object and a contact. At some point, we'll get the genserver to cache this

reset_context(context)

Specs

reset_context(t()) :: t() | nil

Resets the context and sends control back to the parent context if one exists

set_node(context, node)

Specs

set_node(t(), Glific.Flows.Node.t()) :: t()

Set the new node for the context

step_forward(context, body)

Specs

step_forward(t(), String.t()) :: {:ok, map()} | {:error, String.t()}

Given an input string, consume the input and advance the state of the context

update_results(context, key, json)

Specs

update_results(t(), String.t(), map()) :: t()

Update the contact results with each element of the json map

update_results(context, key, input, category)

Specs

update_results(t(), String.t(), String.t(), String.t()) :: t()

Update the contact results state as we step through the flow

wakeup()

Specs

wakeup() :: [t()]

Find all the contexts which need to be woken up and processed