Glific.Groups (Glific v0.9.4)

The Groups context.

Functions

add_permission(query, user)

Specs

add_permission(Ecto.Query.t(), Glific.Users.User.t()) :: Ecto.Query.t()

Add permissioning specific to groups, in this case we want to restrict the visibility of groups that the user can see

change_group(group, attrs \\ %{})

Specs

change_group(Glific.Groups.Group.t(), map()) :: Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking group changes.

Examples

iex> change_group(group)
%Ecto.Changeset{data: %Group{}}

contacts_count(map)

Specs

contacts_count(map()) :: integer()

Return the count of group contacts

count_groups(args)

Specs

count_groups(map()) :: integer()

Return the count of groups, using the same filter as list_groups

create_contact_group(attrs \\ %{})

Specs

create_contact_group(map()) ::
  {:ok, Glific.Groups.ContactGroup.t()} | {:error, Ecto.Changeset.t()}

Creates a contact group.

Examples

iex> create_contact_group(%{field: value})
{:ok, %Contact{}}

iex> create_contact_group(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_group(attrs \\ %{})

Specs

create_group(map()) ::
  {:ok, Glific.Groups.Group.t()} | {:error, Ecto.Changeset.t()}

Creates a group.

Examples

iex> create_group(%{field: value})
{:ok, %Group{}}

iex> create_group(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_user_group(attrs \\ %{})

Specs

create_user_group(map()) ::
  {:ok, Glific.Groups.UserGroup.t()} | {:error, Ecto.Changeset.t()}

Creates a user group.

Examples

iex> create_user_group(%{field: value})
{:ok, %User{}}

iex> create_user_group(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_contact_groups_by_ids(contact_id, group_ids)

Specs

delete_contact_groups_by_ids(integer(), []) :: {integer(), nil | [term()]}

Delete contact groups

delete_group(group)

Specs

delete_group(Glific.Groups.Group.t()) ::
  {:ok, Glific.Groups.Group.t()} | {:error, Ecto.Changeset.t()}

Deletes a group.

Examples

iex> delete_group(group)
{:ok, %Group{}}

iex> delete_group(group)
{:error, %Ecto.Changeset{}}

delete_group_contacts_by_ids(group_id, contact_ids)

Specs

delete_group_contacts_by_ids(integer(), []) :: {integer(), nil | [term()]}

Delete group contacts

delete_group_users_by_ids(group_id, user_ids)

Specs

delete_group_users_by_ids(integer(), []) :: {integer(), nil | [term()]}

Delete group users

delete_user_groups_by_ids(user_id, group_ids)

Specs

delete_user_groups_by_ids(integer(), []) :: {integer(), nil | [term()]}

Delete user groups

get_group!(id)

Specs

get_group!(integer()) :: Glific.Groups.Group.t()

Gets a single group.

Raises Ecto.NoResultsError if the Group does not exist.

Examples

iex> get_group!(123)
%Group{}

iex> get_group!(456)
** (Ecto.NoResultsError)

get_group_ids()

Specs

get_group_ids() :: list()

Fetches all group ids in an organization

list_groups(args, skip_permission \\ false)

Specs

list_groups(map(), boolean()) :: [Glific.Groups.Group.t()]

Returns the list of groups.

Examples

iex> list_groups()
[%Group{}, ...]

update_group(group, attrs)

Specs

update_group(Glific.Groups.Group.t(), map()) ::
  {:ok, Glific.Groups.Group.t()} | {:error, Ecto.Changeset.t()}

Updates a group.

Examples

iex> update_group(group, %{field: new_value})
{:ok, %Group{}}

iex> update_group(group, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_user_groups(map)

Specs

update_user_groups(map()) :: :ok

Updates user groups entries

users_count(map)

Specs

users_count(map()) :: integer()

Return the count of group users