Glific.Messages (Glific v1.1.5)

The Messages context.

Functions

change_message(message, attrs \\ %{})

Specs

change_message(Glific.Messages.Message.t(), map()) :: Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking message changes.

Examples

iex> change_message(message)
%Ecto.Changeset{data: %Message{}}

change_message_media(message_media, attrs \\ %{})

Specs

change_message_media(Glific.Messages.MessageMedia.t(), map()) ::
  Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking message media changes.

Examples

iex> change_message_media(message_media)
%Ecto.Changeset{data: %MessageMedia{}}

clear_messages(contact)

Specs

clear_messages(Glific.Contacts.Contact.t()) :: :ok

Delete all messages of a contact

count_messages(args)

Specs

count_messages(map()) :: integer()

Return the count of messages, using the same filter as list_messages

count_messages_media(args \\ %{})

Specs

count_messages_media(map()) :: integer()

Return the count of messages, using the same filter as list_messages

create_and_send_hsm_message(attrs)

Specs

create_and_send_hsm_message(map()) ::
  {:ok, Glific.Messages.Message.t()} | {:error, String.t()}

Send a hsm template message to the specific contact.

create_and_send_message_to_group(message_params, group)

Specs

create_and_send_message_to_group(map(), Glific.Groups.Group.t()) ::
  {:ok, list()}

Create and send message to all contacts of a group

create_and_send_session_template(template_id, receiver_id)

Specs

create_and_send_session_template(String.t(), integer()) ::
  {:ok, Glific.Messages.Message.t()}

create_and_send_session_template(integer(), integer()) ::
  {:ok, Glific.Messages.Message.t()}

create_and_send_session_template(
  Glific.Templates.SessionTemplate.t() | map(),
  map()
) :: {:ok, Glific.Messages.Message.t()}

Send a session template to the specific contact. This is typically used in automation

create_group_message(attrs)

Specs

create_group_message(map()) ::
  {:ok, Glific.Messages.Message.t()} | {:error, Ecto.Changeset.t()}

Record a message sent to a group in the message table. This message is actually not sent, but is used for display purposes in the group listings

create_message(attrs)

Specs

create_message(map()) ::
  {:ok, Glific.Messages.Message.t()} | {:error, Ecto.Changeset.t()}

Creates a message.

Examples

iex> create_message(%{field: value})
{:ok, %Message{}}

iex> create_message(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_message_media(attrs \\ %{})

Specs

create_message_media(map()) ::
  {:ok, Glific.Messages.MessageMedia.t()} | {:error, Ecto.Changeset.t()}

Creates a message media.

Examples

iex> create_message_media(%{field: value})
{:ok, %MessageMedia{}}

iex> create_message_media(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_temp_message(organization_id, body, attrs \\ [])

Specs

create_temp_message(non_neg_integer(), any(), Keyword.t()) ::
  Glific.Messages.Message.t()

We need to simulate a few messages as we move to the system. This is a wrapper function to add those messages, which trigger specific actions within flows. e.g. include: Completed, Failure, Success etc

delete_message(message)

Specs

delete_message(Glific.Messages.Message.t()) ::
  {:ok, Glific.Messages.Message.t()} | {:error, Ecto.Changeset.t()}

Deletes a message.

Examples

iex> delete_message(message)
{:ok, %Message{}}

iex> delete_message(message)
{:error, %Ecto.Changeset{}}

delete_message_media(message_media)

Specs

delete_message_media(Glific.Messages.MessageMedia.t()) ::
  {:ok, Glific.Messages.MessageMedia.t()} | {:error, Ecto.Changeset.t()}

Deletes a message media.

Examples

iex> delete_message_media(message_media)
{:ok, %MessageMedia{}}

iex> delete_message_media(message_media)
{:error, %Ecto.Changeset{}}

get_message!(id)

Specs

get_message!(integer()) :: Glific.Messages.Message.t()

Gets a single message.

Raises Ecto.NoResultsError if the Message does not exist.

Examples

iex> get_message!(123)
%Message{}

iex> get_message!(456)
** (Ecto.NoResultsError)

get_message_media!(id)

Specs

get_message_media!(integer()) :: Glific.Messages.MessageMedia.t()

Gets a single message media.

Raises Ecto.NoResultsError if the Message media does not exist.

Examples

iex> get_message_media!(123)
%MessageMedia{}

iex> get_message_media!(456)
** (Ecto.NoResultsError)

list_conversations(args, count \\ false)

Specs

list_conversations(map(), boolean()) ::
  [Glific.Conversations.Conversation.t()] | integer()

Given a list of message ids builds a conversation list with most recent conversations at the beginning of the list

list_messages(args)

Specs

list_messages(map()) :: [Glific.Messages.Message.t()]

Returns the list of filtered messages.

Examples

iex> list_messages(map())
[%Message{}, ...]

list_messages_media(args \\ %{})

Specs

list_messages_media(map()) :: [Glific.Messages.MessageMedia.t()]

Returns the list of message media.

Examples

iex> list_messages_media(map())
[%MessageMedia{}, ...]

mark_contact_messages_as_read(contact_id, organization_id)

Specs

mark_contact_messages_as_read(non_neg_integer(), non_neg_integer()) :: nil

Mark that the user has read all messages sent by a given contact

tag_in_message?(message, tag_id)

Specs

tag_in_message?(Glific.Messages.Message.t(), integer()) :: boolean()

Check if the tag is present in message

update_message(message, attrs)

Specs

update_message(Glific.Messages.Message.t(), map()) ::
  {:ok, Glific.Messages.Message.t()} | {:error, Ecto.Changeset.t()}

Updates a message.

Examples

iex> update_message(message, %{field: new_value})
{:ok, %Message{}}

iex> update_message(message, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_message_media(message_media, attrs)

Specs

update_message_media(Glific.Messages.MessageMedia.t(), map()) ::
  {:ok, Glific.Messages.MessageMedia.t()} | {:error, Ecto.Changeset.t()}

Updates a message media.

Examples

iex> update_message_media(message_media, %{field: new_value})
{:ok, %MessageMedia{}}

iex> update_message_media(message_media, %{field: bad_value})
{:error, %Ecto.Changeset{}}