Glific.Partners (Glific v1.3.3)

The Partners context. This is the gateway for the application to access/update all the organization and Provider information.

Functions

active_organizations(orgs)

Specs

active_organizations(list()) :: map()

List of organizations that are active within the system

change_organization(organization, attrs \\ %{})

Specs

change_organization(Glific.Partners.Organization.t(), map()) ::
  Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking organization changes.

Examples

iex> Glific.Partners.change_organization(organization)
%Ecto.Changeset{data: %Glific.Partners.Organization{}}

change_provider(provider, attrs \\ %{})

Specs

change_provider(
  %Glific.Partners.Provider{
    __meta__: term(),
    credential: term(),
    description: term(),
    group: term(),
    id: term(),
    inserted_at: term(),
    is_required: term(),
    keys: term(),
    name: term(),
    organizations: term(),
    secrets: term(),
    shortcode: term(),
    updated_at: term()
  },
  map()
) :: Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking provider changes.

Examples

iex> change_provider(provider)
%Ecto.Changeset{data: %Provider{}}

contact_organization_query(org_id_list)

Specs

contact_organization_query(list()) :: Ecto.Query.t()

Wrapper query used by various statistics collection routines in Glific to return counts on contact with its variations

count_organizations(args \\ %{})

Specs

count_organizations(map()) :: integer()

Return the count of organizations, using the same filter as list_organizations

count_providers(args \\ %{})

Specs

count_providers(map()) :: integer()

Return the count of providers, using the same filter as list_providers

create_credential(attrs)

Specs

create_credential(map()) ::
  {:ok, Glific.Partners.Credential.t()} | {:error, any()}

Creates an organization's credential

create_organization(attrs \\ %{})

Specs

create_organization(map()) ::
  {:ok, Glific.Partners.Organization.t()} | {:error, Ecto.Changeset.t()}

Creates a organization.

Examples

iex> Glific.Partners.create_organization(%{name: value})
{:ok, %Glific.Partners.Organization{}}

iex> Glific.Partners.create_organization(%{bad_field: bad_value})
{:error, %Ecto.Changeset{}}

create_provider(attrs \\ %{})

Specs

create_provider(map()) ::
  {:ok,
   %Glific.Partners.Provider{
     __meta__: term(),
     credential: term(),
     description: term(),
     group: term(),
     id: term(),
     inserted_at: term(),
     is_required: term(),
     keys: term(),
     name: term(),
     organizations: term(),
     secrets: term(),
     shortcode: term(),
     updated_at: term()
   }}
  | {:error, Ecto.Changeset.t()}

Creates a provider.

Examples

iex> create_provider(%{field: value})
{:ok, %Provider{}}

iex> create_provider(%{field: bad_value})
{:error, %Ecto.Changeset{}}

credential_update_callback(organization, arg2)

Specs

credential_update_callback(Glific.Partners.Organization.t(), String.t()) :: :ok

Updating setup

delete_organization(organization)

Specs

delete_organization(Glific.Partners.Organization.t()) ::
  {:ok, Glific.Partners.Organization.t()} | {:error, Ecto.Changeset.t()}

Deletes an Orgsanization.

Examples

iex> Glific.Partners.delete_organization(organization)
{:ok, %Glific.Partners.Organization{}}

iex> delete_organization(organization)
{:error, %Ecto.Changeset{}}

delete_provider(provider)

Specs

delete_provider(%Glific.Partners.Provider{
  __meta__: term(),
  credential: term(),
  description: term(),
  group: term(),
  id: term(),
  inserted_at: term(),
  is_required: term(),
  keys: term(),
  name: term(),
  organizations: term(),
  secrets: term(),
  shortcode: term(),
  updated_at: term()
}) ::
  {:ok,
   %Glific.Partners.Provider{
     __meta__: term(),
     credential: term(),
     description: term(),
     group: term(),
     id: term(),
     inserted_at: term(),
     is_required: term(),
     keys: term(),
     name: term(),
     organizations: term(),
     secrets: term(),
     shortcode: term(),
     updated_at: term()
   }}
  | {:error, Ecto.Changeset.t()}

Deletes a provider.

Examples

iex> delete_provider(provider)
{:ok, %Provider{}}

iex> delete_provider(provider)
{:error, %Ecto.Changeset{}}

disable_credential(organization_id, shortcode)

Specs

disable_credential(non_neg_integer(), String.t()) :: :ok

Disable a specific credential for the organization

fetch_opted_in_contacts(attrs)

Specs

fetch_opted_in_contacts(map()) :: :ok | any()

Fetch opted in contacts data from providers server

fill_cache(organization)

Specs

fill_cache(Glific.Partners.Organization.t()) :: Glific.Partners.Organization.t()

Given a minimal organization object, fill it up and store in cache. Making this public so we can call from test harness and avoid SQL Sandbox issues

get_bsp_balance(organization_id)

Specs

get_bsp_balance(non_neg_integer()) :: {:ok, any()} | {:error, String.t()}

Returns bsp balance for an organization

get_credential(map)

Specs

get_credential(map()) ::
  {:ok, Glific.Partners.Credential.t()} | {:error, String.t() | [String.t()]}

Get organization's credential by service shortcode

get_goth_token(organization_id, provider_shortcode)

Specs

get_goth_token(non_neg_integer(), String.t()) :: nil | Goth.Token.t()

Common function to get the goth config

get_organization!(id)

Specs

get_organization!(integer()) :: Glific.Partners.Organization.t()

Gets a single organization.

Raises Ecto.NoResultsError if the organization does not exist.

Examples

iex> Glific.Partners.get_organization!(1)
%Glific.Partners.Organization{}

iex> Glific.Partners.get_organization!(-1)
** (Ecto.NoResultsError)

get_provider!(id)

Specs

get_provider!(id :: integer()) :: %Glific.Partners.Provider{
  __meta__: term(),
  credential: term(),
  description: term(),
  group: term(),
  id: term(),
  inserted_at: term(),
  is_required: term(),
  keys: term(),
  name: term(),
  organizations: term(),
  secrets: term(),
  shortcode: term(),
  updated_at: term()
}

Gets a single provider.

Raises Ecto.NoResultsError if the Provider does not exist.

Examples

iex> get_provider!(123)
%Provider{}

iex> get_provider!(456)
** (Ecto.NoResultsError)

list_organizations(args \\ %{})

Specs

list_organizations(map()) :: [Glific.Partners.Organization.t()]

Returns the list of organizations.

Examples

iex> Glific.Partners.list_organizations()
[%Glific.Partners.Organization{}, ...]

list_providers(args \\ %{})

Specs

list_providers(map()) :: [
  %Glific.Partners.Provider{
    __meta__: term(),
    credential: term(),
    description: term(),
    group: term(),
    id: term(),
    inserted_at: term(),
    is_required: term(),
    keys: term(),
    name: term(),
    organizations: term(),
    secrets: term(),
    shortcode: term(),
    updated_at: term()
  },
  ...
]

Returns the list of providers.

Examples

iex> list_providers()
[%Provider{}, ...]

load_cache(cachex_key)

Specs

load_cache(tuple()) :: {:ignore, Glific.Partners.Organization.t()}

Follow the cachex protocol to load the cache from the DB

org_id_list(list, recent)

Specs

org_id_list(list(), boolean()) :: list()

Given an empty list, determine which organizations have been active in the recent past

organization(cache_key)

Specs

organization(non_neg_integer() | String.t()) ::
  Glific.Partners.Organization.t() | nil | {:error, String.t()}

Cache the entire organization structure.

organization_contact_id(organization_id)

Specs

organization_contact_id(non_neg_integer()) :: integer()

This contact id is special since it is the sender for all outbound messages and the receiver for all inbound messages

organization_language_id(organization_id)

Specs

organization_language_id(non_neg_integer()) :: integer()

Get the default language id

organization_timezone(organization_id)

Specs

organization_timezone(non_neg_integer()) :: String.t()

Get the timezone

perform_all(handler, handler_args, list, only_recent \\ false)

Specs

perform_all((... -> nil), map() | nil, list() | [] | nil, boolean()) :: :ok

Execute a function across all active organizations. This function is typically called by a cron job worker process

The handler is expected to take the organization id as its first argument. The second argument is expected to be a map of arguments passed in by the cron job, and can be ignored if not used

The list is a restricted list of organizations, so we dont repeatedly do work. The convention is as follows:

list == nil - the action should not be performed for any organization list == [] (empty list) - the action should be performed for all organizations list == [ values ] - the actions should be performed only for organizations in the values list

recent_organizations(map, bool)

Specs

recent_organizations(map(), boolean()) :: map()

Get the organizations which had a message transaction in the last minutes as defined by @active_minutes

remove_organization_cache(organization_id, shortcode)

Specs

remove_organization_cache(non_neg_integer(), String.t()) :: any()

Removing organization cache

update_credential(credential, attrs)

Specs

update_credential(Glific.Partners.Credential.t(), map()) ::
  {:ok, Glific.Partners.Credential.t()} | {:error, Ecto.Changeset.t()}

Updates an organization's credential

update_organization(organization, attrs)

Specs

update_organization(Glific.Partners.Organization.t(), map()) ::
  {:ok, Glific.Partners.Organization.t()} | {:error, Ecto.Changeset.t()}

Updates an organization.

Examples

iex> Glific.Partners.update_organization(Organization, %{name: new_name})
{:ok, %Glific.Partners.Organization{}}

iex> Glific.Partners.update_organization(Organization, %{abc: bad_value})
{:error, %Ecto.Changeset{}}

update_provider(provider, attrs)

Specs

update_provider(
  %Glific.Partners.Provider{
    __meta__: term(),
    credential: term(),
    description: term(),
    group: term(),
    id: term(),
    inserted_at: term(),
    is_required: term(),
    keys: term(),
    name: term(),
    organizations: term(),
    secrets: term(),
    shortcode: term(),
    updated_at: term()
  },
  map()
) ::
  {:ok,
   %Glific.Partners.Provider{
     __meta__: term(),
     credential: term(),
     description: term(),
     group: term(),
     id: term(),
     inserted_at: term(),
     is_required: term(),
     keys: term(),
     name: term(),
     organizations: term(),
     secrets: term(),
     shortcode: term(),
     updated_at: term()
   }}
  | {:error, Ecto.Changeset.t()}

Updates a provider.

Examples

iex> update_provider(provider, %{field: new_value})
{:ok, %Provider{}}

iex> update_provider(provider, %{field: bad_value})
{:error, %Ecto.Changeset{}}