Glific v0.2.0 Glific.Partners

The Partners context. This is the gateway for the application to access/update all the organization and Provider information.

Functions

change_organization(organization, attrs \\ %{})

Specs

change_organization(Glific.Partners.Organization.t(), map()) ::
  Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking organization changes.

Examples

iex> Glific.Partners.change_organization(organization)
%Ecto.Changeset{data: %Glific.Partners.Organization{}}

change_provider(provider, attrs \\ %{})

Specs

change_provider(
  %Glific.Partners.Provider{
    __meta__: term(),
    api_end_point: term(),
    id: term(),
    inserted_at: term(),
    name: term(),
    organizations: term(),
    updated_at: term(),
    url: term()
  },
  map()
) :: Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking provider changes.

Examples

iex> change_provider(provider)
%Ecto.Changeset{data: %Provider{}}

count_organizations(args \\ %{})

Specs

count_organizations(map()) :: integer()

Return the count of organizations, using the same filter as list_organizations

count_providers(args \\ %{})

Specs

count_providers(map()) :: integer()

Return the count of providers, using the same filter as list_providers

create_organization(attrs \\ %{})

Specs

create_organization(map()) ::
  {:ok, Glific.Partners.Organization.t()} | {:error, Ecto.Changeset.t()}

Creates a organization.

Examples

iex> Glific.Partners.create_organization(%{name: value})
{:ok, %Glific.Partners.Organization{}}

iex> Glific.Partners.create_organization(%{bad_field: bad_value})
{:error, %Ecto.Changeset{}}

create_provider(attrs \\ %{})

Specs

create_provider(map()) ::
  {:ok,
   %Glific.Partners.Provider{
     __meta__: term(),
     api_end_point: term(),
     id: term(),
     inserted_at: term(),
     name: term(),
     organizations: term(),
     updated_at: term(),
     url: term()
   }}
  | {:error, Ecto.Changeset.t()}

Creates a provider.

Examples

iex> create_provider(%{field: value})
{:ok, %Provider{}}

iex> create_provider(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_organization(organization)

Specs

delete_organization(Glific.Partners.Organization.t()) ::
  {:ok, Glific.Partners.Organization.t()} | {:error, Ecto.Changeset.t()}

Deletes an Organization.

Examples

iex> Glific.Partners.delete_organization(organization)
{:ok, %Glific.Partners.Organization{}}

iex> delete_organization(organization)
{:error, %Ecto.Changeset{}}

delete_provider(provider)

Specs

delete_provider(%Glific.Partners.Provider{
  __meta__: term(),
  api_end_point: term(),
  id: term(),
  inserted_at: term(),
  name: term(),
  organizations: term(),
  updated_at: term(),
  url: term()
}) ::
  {:ok,
   %Glific.Partners.Provider{
     __meta__: term(),
     api_end_point: term(),
     id: term(),
     inserted_at: term(),
     name: term(),
     organizations: term(),
     updated_at: term(),
     url: term()
   }}
  | {:error, Ecto.Changeset.t()}

Deletes a provider.

Examples

iex> delete_provider(provider)
{:ok, %Provider{}}

iex> delete_provider(provider)
{:error, %Ecto.Changeset{}}

get_organization!(id)

Specs

get_organization!(integer()) :: Glific.Partners.Organization.t()

Gets a single organization.

Raises Ecto.NoResultsError if the organization does not exist.

Examples

iex> Glific.Partners.get_organization!(1)
%Glific.Partners.Organization{}

iex> Glific.Partners.get_organization!(-1)
** (Ecto.NoResultsError)

get_provider!(id)

Specs

get_provider!(id :: integer()) :: %Glific.Partners.Provider{
  __meta__: term(),
  api_end_point: term(),
  id: term(),
  inserted_at: term(),
  name: term(),
  organizations: term(),
  updated_at: term(),
  url: term()
}

Gets a single provider.

Raises Ecto.NoResultsError if the Provider does not exist.

Examples

iex> get_provider!(123)
%Provider{}

iex> get_provider!(456)
** (Ecto.NoResultsError)

list_organizations(args \\ %{})

Specs

list_organizations(map()) :: [Glific.Partners.Organization.t()]

Returns the list of organizations.

Examples

iex> Glific.Partners.list_organizations()
[%Glific.Partners.Organization{}, ...]

list_providers(args \\ %{})

Specs

list_providers(map()) :: [
  %Glific.Partners.Provider{
    __meta__: term(),
    api_end_point: term(),
    id: term(),
    inserted_at: term(),
    name: term(),
    organizations: term(),
    updated_at: term(),
    url: term()
  },
  ...
]

Returns the list of providers.

Examples

iex> list_providers()
[%Provider{}, ...]

organization_contact_id()

Specs

organization_contact_id() :: integer()

We will cache this soon, since this is a frequently requested item. This contact id is special since it is the sender for all outbound messages and the receiver for all inbound messages

update_organization(provider, attrs)

Specs

update_organization(Glific.Partners.Organization.t(), map()) ::
  {:ok, Glific.Partners.Organization.t()} | {:error, Ecto.Changeset.t()}

Updates an organization.

Examples

iex> Glific.Partners.update_organization(Organization, %{name: new_name})
{:ok, %Glific.Partners.Organization{}}

iex> Glific.Partners.update_organization(Organization, %{abc: bad_value})
{:error, %Ecto.Changeset{}}

update_provider(provider, attrs)

Specs

update_provider(
  %Glific.Partners.Provider{
    __meta__: term(),
    api_end_point: term(),
    id: term(),
    inserted_at: term(),
    name: term(),
    organizations: term(),
    updated_at: term(),
    url: term()
  },
  map()
) ::
  {:ok,
   %Glific.Partners.Provider{
     __meta__: term(),
     api_end_point: term(),
     id: term(),
     inserted_at: term(),
     name: term(),
     organizations: term(),
     updated_at: term(),
     url: term()
   }}
  | {:error, Ecto.Changeset.t()}

Updates a provider.

Examples

iex> update_provider(provider, %{field: new_value})
{:ok, %Provider{}}

iex> update_provider(provider, %{field: bad_value})
{:error, %Ecto.Changeset{}}