Glific.Partners.Billing (Glific v4.3.0)

We will use this as the main context interface for all billing subscriptions and the stripe interface.

Types

t()

@type t() :: %Glific.Partners.Billing{
  __meta__: Ecto.Schema.Metadata.t(),
  currency: String.t() | nil,
  deduct_tds: boolean() | false,
  email: String.t() | nil,
  id: non_neg_integer() | nil,
  inserted_at: :utc_datetime | nil,
  is_active: boolean() | true,
  is_delinquent: boolean(),
  name: String.t() | nil,
  organization: term(),
  organization_id: term(),
  stripe_current_period_end: DateTime.t() | nil,
  stripe_current_period_start: DateTime.t() | nil,
  stripe_customer_id: String.t() | nil,
  stripe_last_usage_recorded: DateTime.t() | nil,
  stripe_payment_method_id: String.t() | nil,
  stripe_subscription_id: String.t() | nil,
  stripe_subscription_items: map(),
  stripe_subscription_status: String.t() | nil,
  tds_amount: float() | nil,
  updated_at: :utc_datetime | nil
}

Functions

changeset(billing, attrs)

@spec changeset(t(), map()) :: Ecto.Changeset.t()

Standard changeset pattern we use for all data types

create(organization, attrs)

@spec create(Glific.Partners.Organization.t(), map()) ::
  {:ok, t()} | {:error, Ecto.Changeset.t() | String.t()}

Create a billing record in glific, a billing customer in Stripe, given an organization

create_billing(attrs \\ %{})

@spec create_billing(map()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Create a billing record

create_subscription(organization, params)

@spec create_subscription(Glific.Partners.Organization.t(), map()) ::
  {:ok, Stripe.Subscription.t()} | {:pending, map()} | {:error, String.t()}

Once the organization has entered a new payment card we create a subscription for it. We'll do updating the card in a seperate function

credit_customer(transaction)

@spec credit_customer(map()) :: any() | non_neg_integer()

Adding credit to customer in Stripe

customer_portal_link(billing)

@spec customer_portal_link(t()) :: {:ok, any()} | {:error, String.t()}

fetches customer portal url of organization with billing status as active

delete_billing(billing)

@spec delete_billing(t()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Delete the billing record

get_billing(clauses)

@spec get_billing(map()) :: t() | nil

Retrieve a billing record by clauses

get_promo_codes(code)

@spec get_promo_codes(any()) :: any()

Validate entered coupon code and return with coupon details

make_stripe_request(endpoint, method, params, opts \\ [])

@spec make_stripe_request(String.t(), atom(), map(), list()) :: any()

A common function for making Stripe API calls with params that are not supported withing Stripity Stripe

period_usage(record_date)

@spec period_usage(DateTime.t()) :: :ok

This is called on a regular schedule to update usage.

record_usage(organization_id, start_date, end_date)

@spec record_usage(non_neg_integer(), DateTime.t(), DateTime.t()) :: :ok

Record the usage for a specific organization from start_date to end_date both dates inclusive

stripe_ids()

@spec stripe_ids() :: map()

Fetch the stripe id's

subscription_created_callback(subscription, org_id)

@spec subscription_created_callback(Stripe.Subscription.t(), non_neg_integer()) ::
  :ok | {:error, Stripe.Error.t()}

Stripe subscription created callback via webhooks. We are using this to update the prorate data with monthly billing.

tax_rates()

@spec tax_rates() :: list()

Fetch the stripe tax rates

update_billing(billing, attrs)

@spec update_billing(t(), map()) :: {:ok, t()} | {:error, Ecto.Changeset.t()}

Upate the billing record

update_payment_method(organization, stripe_payment_method_id)

@spec update_payment_method(Glific.Partners.Organization.t(), String.t()) ::
  {:ok, t()} | {:error, map()}

Update organization and stripe customer with the current payment method as returned by stripe

update_stripe_customer(billing, attrs)

@spec update_stripe_customer(t(), map()) :: {:ok, t()} | {:error, Stripe.Error.t()}

Upate the stripe customer details record

update_subscription(billing, organization)

@spec update_subscription(t(), Glific.Partners.Organization.t()) ::
  Glific.Partners.Organization.t()

Update organization subscription plan

update_subscription_details(subscription, organization_id, billing)

@spec update_subscription_details(
  Stripe.Subscription.t(),
  non_neg_integer(),
  t() | nil
) ::
  {:ok, Stripe.Subscription.t()} | {:error, String.t()}

Update subscription details. We will also use this method while updating the details form webhook.

update_usage(organization_id, map)

@spec update_usage(non_neg_integer(), map()) :: :ok

Update the usage record for all active subscriptions on a daily and weekly basis