Glific.Seeds.SeedsMigration (Glific v1.3.3)

One shot migration of data to add simulators and saas admin. We use the functions in this file to add simulators for new organizations as they are created

Functions

add_saas_user(organization, language)

Specs

add_saas_user(Glific.Partners.Organization.t(), Glific.Settings.Language.t()) ::
  :ok

Add a saas user for the organization. We need to check if it already exists since this code is used during data migration and can be repeated for the same organization

fix_message_number(org_id)

Specs

fix_message_number(list() | integer()) :: :ok

Reset message number for a list of organizations or for a org_id

migrate_data(phase, organization \\ nil)

Specs

migrate_data(atom(), Glific.Partners.Organization.t() | nil) :: :ok

Public interface to do a seed migration across all organizations.

One function to rule them all. This function is invoked manually by a glific developer to add data from the DB. This seems the cleanest way to do such things. We use phases to seperate different migrations

sync_schema_with_bigquery(org_id_list)

Specs

sync_schema_with_bigquery(list()) :: :ok

Sync bigquery schema with local db changes.