Glific v0.2.0 Glific.Settings

The Settings context. This includes language for now.

Functions

change_language(language, attrs \\ %{})

Specs

change_language(Glific.Settings.Language.t(), map()) :: Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking language changes.

Examples

iex> change_language(language)
%Ecto.Changeset{data: %Language{}}

count_languages(args \\ %{})

Specs

count_languages(map()) :: integer()

Return the count of languages, using the same filter as list_languages

create_language(attrs \\ %{})

Specs

create_language(map()) ::
  {:ok, Glific.Settings.Language.t()} | {:error, Ecto.Changeset.t()}

Creates a language.

Examples

iex> create_language(%{field: value})
{:ok, %Language{}}

iex> create_language(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_language(language)

Specs

delete_language(Glific.Settings.Language.t()) ::
  {:ok, Glific.Settings.Language.t()} | {:error, Ecto.Changeset.t()}

Deletes a language.

Examples

iex> delete_language(language)
{:ok, %Language{}}

iex> delete_language(language)
{:error, %Ecto.Changeset{}}

get_language!(id)

Specs

get_language!(integer()) :: Glific.Settings.Language.t()

Gets a single language.

Raises Ecto.NoResultsError if the Language does not exist.

Examples

iex> get_language!(123)
%Language{}

iex> get_language!(456)
** (Ecto.NoResultsError)

language_upsert(attrs)

Specs

language_upsert(map()) :: {:ok, Glific.Settings.Language.t()}

Upserts a language based on the unique indexes in the table. If there is a match it returns the existing contact, else it creates a new one

list_languages(args \\ %{})

Specs

list_languages(map()) :: [Glific.Settings.Language.t(), ...]

Returns the list of languages.

Examples

iex> list_languages()
[%Language{}, ...]

locale_id_map()

Specs

locale_id_map() :: %{required(String.t()) => integer()}

Get map of localte to ids for easier lookup for json based flow editor

update_language(language, attrs)

Specs

update_language(Glific.Settings.Language.t(), map()) ::
  {:ok, Glific.Settings.Language.t()} | {:error, Ecto.Changeset.t()}

Updates a language.

Examples

iex> update_language(language, %{field: new_value})
{:ok, %Language{}}

iex> update_language(language, %{field: bad_value})
{:error, %Ecto.Changeset{}}