Glific v0.1.0 Glific.Tags

The Tags Context, which encapsulates and manages tags and the related join tables.

Functions

change_contact_tag(contact_tag, attrs \\ %{})

Specs

change_contact_tag(Glific.Tags.ContactTag.t(), map()) :: Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking contact changes.

Examples

iex> change_contact_tag(contact_tag)
%Ecto.Changeset{data: %ContactTag{}}

change_message_tag(message_tag, attrs \\ %{})

Specs

change_message_tag(Glific.Tags.MessageTag.t(), map()) :: Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking message changes.

Examples

iex> change_message_tag(message_tag)
%Ecto.Changeset{data: %MessageTag{}}

change_tag(tag, attrs \\ %{})

Specs

change_tag(Glific.Tags.Tag.t(), map()) :: Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking tag changes.

Examples

iex> change_tag(tag)
%Ecto.Changeset{data: %Tag{}}

count_tags(args \\ %{})

Specs

count_tags(map()) :: integer()

Return the count of tags, using the same filter as list_tags

create_contact_tag(attrs \\ %{})

Specs

create_contact_tag(map()) ::
  {:ok, Glific.Tags.ContactTag.t()} | {:error, Ecto.Changeset.t()}

Creates a contact.

Examples

iex> create_contact_tag(%{field: value})
{:ok, %Contact{}}

iex> create_contact_tag(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_message_tag(attrs \\ %{})

Specs

create_message_tag(map()) ::
  {:ok, Glific.Tags.MessageTag.t()} | {:error, Ecto.Changeset.t()}

Creates a message tag

Examples

iex> create_message_tag(%{field: value}) {:ok, %Message{}}

iex> create_message_tag(%{field: bad_value}) {:error, %Ecto.Changeset{}}

create_tag(attrs \\ %{})

Specs

create_tag(map()) :: {:ok, Glific.Tags.Tag.t()} | {:error, Ecto.Changeset.t()}

Creates a tag.

Examples

iex> create_tag(%{field: value})
{:ok, %Tag{}}

iex> create_tag(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_contact_tag(contact_tag)

Specs

delete_contact_tag(Glific.Tags.ContactTag.t()) ::
  {:ok, Glific.Tags.ContactTag.t()} | {:error, Ecto.Changeset.t()}

Deletes a contact tag.

Examples

iex> delete_contact_tag(contact_tag)
{:ok, %ContactTag{}}

iex> delete_contact_tag(contact_tag)
{:error, %Ecto.Changeset{}}

delete_message_tag(message_tag)

Specs

delete_message_tag(Glific.Tags.MessageTag.t()) ::
  {:ok, Glific.Tags.MessageTag.t()} | {:error, Ecto.Changeset.t()}

Deletes a message tag.

Examples

iex> delete_message_tag(message_tag)
{:ok, %MessageTag{}}

iex> delete_message_tag(message_tag)
{:error, %Ecto.Changeset{}}

delete_message_tag_by_ids(message_id, tag_id)

Specs

delete_message_tag_by_ids(integer(), integer()) :: {integer(), nil | [term()]}

delete_message_tag_by_ids(integer(), []) :: {integer(), nil | [term()]}

In Join tables we rarely use the table id. We always know the object ids and hence more convenient to delete an entry via its object ids. We will generalize this function and move it to Repo.ex when we get a better handle on how to do so :)

delete_tag(tag)

Specs

delete_tag(Glific.Tags.Tag.t()) ::
  {:ok, Glific.Tags.Tag.t()} | {:error, Ecto.Changeset.t()}

Deletes a tag.

Examples

iex> delete_tag(tag)
{:ok, %Tag{}}

iex> delete_tag(tag)
{:error, %Ecto.Changeset{}}

get_contact_tag!(id)

Specs

get_contact_tag!(integer()) :: Glific.Tags.ContactTag.t()

Gets a single contact.

Raises Ecto.NoResultsError if the Contact does not exist.

Examples

iex> get_contact_tag!(123)
%Contact{}

iex> get_contact_tag!(456)
** (Ecto.NoResultsError)

get_message_tag!(id)

Specs

get_message_tag!(integer()) :: Glific.Tags.MessageTag.t()

Gets a single message.

Raises Ecto.NoResultsError if the Message does not exist.

Examples

iex> get_message_tag!(123)
%Message{}

iex> get_message_tag!(456)
** (Ecto.NoResultsError)

get_tag!(id)

Specs

get_tag!(integer()) :: Glific.Tags.Tag.t()

Gets a single tag.

Raises Ecto.NoResultsError if the Tag does not exist.

Examples

iex> get_tag!(123)
%Tag{}

iex> get_tag!(456)
** (Ecto.NoResultsError)

keyword_map()

Specs

keyword_map() :: map()

Converts all tag kewords into the map where keyword is the key and tag id is the value

list_contacts_tags(args \\ %{})

Specs

list_contacts_tags(map()) :: [Glific.Tags.ContactTag.t()]

Returns the list of contacts tags.

Examples

iex> list_contacts_tags()
[%ContactTag{}, ...]

list_messages_tags(args \\ %{})

Specs

list_messages_tags(map()) :: [Glific.Tags.MessageTag.t()]

Returns the list of messages tags.

Examples

iex> list_messages_tags()
[%MessageTag{}, ...]

list_tags(args \\ %{})

Specs

list_tags(map()) :: [Glific.Tags.Tag.t()]

Returns the list of tags.

Examples

iex> list_tags()
[%Tag{}, ...]

remove_tag_from_all_message(contact_id, tag_label)

Specs

remove_tag_from_all_message(integer(), String.t()) :: list()

Remove a specific tag from contact messages

status_map()

Specs

status_map() :: %{required(String.t()) => integer()}

Filter all the status tag and returns as a map

tags_map(tags)

Specs

tags_map([String.t()]) :: %{required(String.t()) => integer()}

Generic function to build a tag map for easy queries. Suspect we'll need it for all objects soon, and will promote this to Repo

update_contact_tag(contact_tag, attrs)

Specs

update_contact_tag(Glific.Tags.ContactTag.t(), map()) ::
  {:ok, Glific.Tags.ContactTag.t()} | {:error, Ecto.Changeset.t()}

Updates a contact tag.

Examples

iex> update_contact_tag(contact_tag, %{field: new_value})
{:ok, %ContactTag{}}

iex> update_contact_tag(contact_tag, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_message_tag(message_tag, attrs)

Specs

update_message_tag(Glific.Tags.MessageTag.t(), map()) ::
  {:ok, Glific.Tags.MessageTag.t()} | {:error, Ecto.Changeset.t()}

Updates a message tag.

Examples

iex> update_message_tag(message_tag, %{field: new_value})
{:ok, %MessageTag{}}

iex> update_message_tag(message_tag, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_tag(tag, attrs)

Specs

update_tag(Glific.Tags.Tag.t(), map()) ::
  {:ok, Glific.Tags.Tag.t()} | {:error, Ecto.Changeset.t()}

Updates a tag.

Examples

iex> update_tag(tag, %{field: new_value})
{:ok, %Tag{}}

iex> update_tag(tag, %{field: bad_value})
{:error, %Ecto.Changeset{}}