Glific v0.1.0 Glific.Templates

The Templates context.

Functions

change_session_template(session_template, attrs \\ %{})

Specs

change_session_template(Glific.Templates.SessionTemplate.t(), map()) ::
  Ecto.Changeset.t()

Returns an %Ecto.Changeset{} for tracking session_template changes.

Examples

iex> change_session_template(session_template)
%Ecto.Changeset{data: %SessionTemplate{}}

count_session_templates(args \\ %{})

Specs

count_session_templates(map()) :: integer()

Return the count of session_templates, using the same filter as list_session_templates

create_session_template(attrs \\ %{})

Specs

create_session_template(map()) ::
  {:ok, Glific.Templates.SessionTemplate.t()} | {:error, Ecto.Changeset.t()}

Creates a session_template.

Examples

iex> create_session_template(%{field: value})
{:ok, %SessionTemplate{}}

iex> create_session_template(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_template_from_message(map)

Specs

create_template_from_message(%{message_id: integer(), input: map()}) ::
  {:ok, Glific.Templates.SessionTemplate.t()} | {:error, String.t()}

Create a session template form message Body and type will be the message attributes

delete_session_template(session_template)

Specs

delete_session_template(Glific.Templates.SessionTemplate.t()) ::
  {:ok, Glific.Templates.SessionTemplate.t()} | {:error, Ecto.Changeset.t()}

Deletes a session_template.

Examples

iex> delete_session_template(session_template)
{:ok, %SessionTemplate{}}

iex> delete_session_template(session_template)
{:error, %Ecto.Changeset{}}

get_session_template!(id)

Specs

get_session_template!(integer()) :: Glific.Templates.SessionTemplate.t()

Gets a single session_template.

Raises Ecto.NoResultsError if the SessionTemplate does not exist.

Examples

iex> get_session_template!(123)
%SessionTemplate{}

iex> get_session_template!(456)
** (Ecto.NoResultsError)

list_session_templates(args \\ %{})

Specs

list_session_templates(map()) :: [Glific.Templates.SessionTemplate.t()]

Returns the list of session_templates.

Examples

iex> list_session_templates()
[%SessionTemplate{}, ...]

template_upsert(attrs)

Specs

template_upsert(map()) :: {:ok, Glific.Templates.SessionTemplate.t()}

Gets or Creates a template based on the unique indexes in the table. If there is a match it returns the existing template, else it creates a new one

update_session_template(session_template, attrs)

Specs

update_session_template(Glific.Templates.SessionTemplate.t(), map()) ::
  {:ok, Glific.Templates.SessionTemplate.t()} | {:error, Ecto.Changeset.t()}

Updates a session_template.

Examples

iex> update_session_template(session_template, %{field: new_value})
{:ok, %SessionTemplate{}}

iex> update_session_template(session_template, %{field: bad_value})
{:error, %Ecto.Changeset{}}