GlificWeb.Flows.FlowEditorController (Glific v0.9.4)

The Flow Editor Controller

Functions

activity(conn, params)

Specs

activity(Plug.Conn.t(), nil | maybe_improper_list() | map()) :: Plug.Conn.t()

This is used to checking if the connection between frontend and backend is established or not.

channels(conn, params)

Specs

channels(Plug.Conn.t(), nil | maybe_improper_list() | map()) :: Plug.Conn.t()

A list of all the communication channels. For Glific it's just WhatsApp. We are not supporting them for now. We will come back to this in near future

classifiers(conn, params)

Specs

classifiers(Plug.Conn.t(), nil | maybe_improper_list() | map()) :: Plug.Conn.t()

A list of all the NLP classifiers. For Glific it's just WhatsApp. We are not supporting them for now. We will come back to this in near future

completion(conn, params)

Specs

completion(Plug.Conn.t(), nil | maybe_improper_list() | map()) :: Plug.Conn.t()

instead of reading a file we can call it directly from Assets. We will come back on that when we have more clearity of the use cases

fields_post(conn, params)

Specs

fields_post(Plug.Conn.t(), map()) :: Plug.Conn.t()

Add Contact fields into the database. The response should be a map with 3 keys % { Key: Field name, name: Field display name value_type: type of the value}

We are not supporting this for now. We will add that in future

flows(conn, map)

Specs

flows(Plug.Conn.t(), nil | maybe_improper_list() | map()) :: Plug.Conn.t()

Let's get all the flows or a latest flow revision

functions(conn, _)

Specs

functions(Plug.Conn.t(), nil | maybe_improper_list() | map()) :: Plug.Conn.t()

all the supported funcations we provide

labels(conn, params)

Specs

labels(Plug.Conn.t(), nil | maybe_improper_list() | map()) :: Plug.Conn.t()

Get all the tags so that user can apply them on incoming message. We are not supporting this for now. To enable It should return a list of map having uuid and name as keys [%{uuid: tag.uuid, name: tag.label}]

labels_post(conn, params)

Specs

labels_post(Plug.Conn.t(), nil | maybe_improper_list() | map()) :: Plug.Conn.t()

Store a label (new tag) in the system. The return response should be a map of 3 keys. [%{uuid: tag.uuid, name: params["name"], count}]

We are not supporting them for now. We will come back to this in near future

resthooks(conn, params)

Specs

resthooks(Plug.Conn.t(), nil | maybe_improper_list() | map()) :: Plug.Conn.t()

We are not using this for now but this is required for flow editor config.

revisions(conn, map)

Specs

revisions(Plug.Conn.t(), nil | maybe_improper_list() | map()) :: Plug.Conn.t()

Get all or a specific revision for a flow

save_revisions(conn, params)

Specs

save_revisions(Plug.Conn.t(), nil | maybe_improper_list() | map()) ::
  Plug.Conn.t()

Save a revision for a flow and get the revision id

ticketers(conn, params)

Specs

ticketers(Plug.Conn.t(), nil | maybe_improper_list() | map()) :: Plug.Conn.t()

We are not sure how to use this but this endpoint is required for flow editor. Will come back to this in future.

validate_media(conn, params)

Specs

validate_media(Plug.Conn.t(), nil | maybe_improper_list() | map()) ::
  Plug.Conn.t()

Validate media to send as attachment