GlificWeb.Plug (Glific v0.9.4)

This file and the below files have been "borrowed and modified" from triplex: https://github.com/ateliware/triplex The original copyright and license - MIT belong to the authors and contributors of Triplex

plug.ex

ensure_plug.ex

ensure_plug_config.ex

subdomain_plug.ex

subdomain_plug_config.ex

param_plug.ex

param_plug_config.ex

This module have some basic functions for our triplex plugs.

The plugs we have for now are:

GlificWeb.ParamPlug - loads the organization from a body or query param

GlificWeb.SubdomainPlug - loads the organization from the url subdomain

GlificWeb.EnsurePlug - ensures the current organization is loaded and halts if not

Functions

ensure_organization(conn, config)

Specs

ensure_organization(Plug.Conn.t(), map()) :: Plug.Conn.t()

Ensure the organization is loaded, and if not, halts the conn.

The config map/struct must have:

assign: the name of the assign where we must save the organization.

put_organization(conn, organization, config)

Specs

put_organization(Plug.Conn.t(), any(), map()) :: Plug.Conn.t()

Puts the given organization as an assign on the given conn, but only if the organization is not reserved.

The config map/struct must have:

organization_handler: function to handle the organization param. Its return will be used as the organization.

assign: the name of the assign where we must save the organization.