GlificWeb.Providers.Gupshup.Controllers.MessageEventController (Glific v0.9.4)

Dedicated controller to handle all the message status requests like read, delivered etc..

Functions

delivered(conn, params)

Specs

delivered(Plug.Conn.t(), map()) :: Plug.Conn.t()

Message has been delivered to whats app

enqueued(conn, params)

Specs

enqueued(Plug.Conn.t(), map()) :: Plug.Conn.t()

Message status when the message has been sent to gupshup

failed(conn, params)

Specs

failed(Plug.Conn.t(), map()) :: Plug.Conn.t()

Message status when gupshup could not send the message to whats app

handler(conn, params, msg)

Specs

handler(Plug.Conn.t(), map(), String.t()) :: Plug.Conn.t()

Default handle for all message event callbacks

read(conn, params)

Specs

read(Plug.Conn.t(), map()) :: Plug.Conn.t()

Message has been read by the beneficiary

sent(conn, params)

Specs

sent(Plug.Conn.t(), map()) :: Plug.Conn.t()

Message has been sent to whats app via Gupshup