GlificWeb.Resolvers.Flows (Glific v0.9.4)

Flow Resolver which sits between the GraphQL schema and Glific Flow Context API. This layer basically stiches together one or more calls to resolve the incoming queries.

Functions

copy_flow(_, map1, map2)

Specs

copy_flow(Absinthe.Resolution.t(), %{id: integer(), input: map()}, %{
  context: map()
}) :: {:ok, any()} | {:error, any()}

Make a copy a flow

count_flows(_, args, _)

Specs

count_flows(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, integer()}

Get the count of flows filtered by args

flow(_, map1, map2)

Specs

flow(Absinthe.Resolution.t(), %{id: integer()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Get a specific flow by id

flows(_, args, _)

Specs

flows(Absinthe.Resolution.t(), map(), %{context: map()}) ::
  {:ok, [Glific.Flows.Flow]}

Get the list of flows

publish_flow(_, map, _)

Specs

publish_flow(Absinthe.Resolution.t(), %{uuid: String.t()}, %{context: map()}) ::
  {:ok, any()} | {:error, any()}

Publish a flow

start_contact_flow(_, map1, map2)

Specs

start_contact_flow(
  Absinthe.Resolution.t(),
  %{flow_id: integer(), contact_id: integer()},
  %{context: map()}
) :: {:ok, any()} | {:error, any()}

Start a flow for a contact

start_group_flow(_, map1, map2)

Specs

start_group_flow(
  Absinthe.Resolution.t(),
  %{flow_id: integer(), group_id: integer()},
  %{context: map()}
) :: {:ok, any()} | {:error, any()}

Start a flow for all contacts of a group