Glific v0.1.0 GlificWeb.Router.Helpers

Module with named helpers generated from GlificWeb.Router.

Functions

api_v1_registration_path(conn_or_endpoint, action)

api_v1_registration_path(conn_or_endpoint, action, params)

api_v1_registration_url(conn_or_endpoint, action)

api_v1_registration_url(conn_or_endpoint, action, params)

api_v1_session_path(conn_or_endpoint, action)

api_v1_session_path(conn_or_endpoint, action, params)

api_v1_session_url(conn_or_endpoint, action)

api_v1_session_url(conn_or_endpoint, action, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.