GlificWeb.Schema (Glific v0.9.4)

This is the container for the top level Absinthe GraphQL schema which encapsulates the entire Glific Public API. This file is primarily a container and pulls in the relevant information for data type specific files.

Functions

config_fun(args, map)

Specs

config_fun(map(), map()) :: {:ok, Keyword.t()} | {:error, String.t()}

Validation function for all subscriptions received by the system

context(context)

Specs

context(map()) :: map()

Used to set some values in the context that we may need in order to run. We store the organization id and the current user in the context once the user has been authenticated

middleware(middleware, field, arg3)

Specs

middleware(
  [Absinthe.Middleware.spec(), ...],
  Absinthe.Type.Field.t(),
  Absinthe.Type.Object.t()
) :: [Absinthe.Middleware.spec(), ...]

Used to apply middleware on all or a group of fields based on pattern matching.

It is passed the existing middleware for a field, the field itself, and the object that the field is a part of.

plugins()

Specs

plugins() :: [Absinthe.Plugin.t()]

Used to define the list of plugins to run before and after resolution.